Tool registry stage 2: convert the 34 member-tier tools to ToolDef domains
PR B of docs/TOOL-REGISTRY-DESIGN.md's strangler migration — every member-tier
closure tool moves out of buildToolServer into a declarative domain file,
bodies byte-verbatim, metadata mirrored from rbac.ts/core.ts (cross-checked by
tests/toolRegistry.test.ts), module state moved exactly once:

- src/agent/tools/info.ts: community_info, community_guidelines, check_status,
  list_events (platforms:['discord'] per DISCORD_ONLY_TOOLS) + the four static
  capability rundown texts + EVENTS_LIST_LIMIT.
- src/agent/tools/knowledgeMember.ts: knowledge_search (turnState gap/stale
  correlation writes intact), list_knowledge_topics, suggest_knowledge,
  withdraw_knowledge_tip.
- src/agent/tools/memory.ts: remember_search, catch_up +
  RECALL_TRUNCATION_CHARS and the CATCH_UP_* constants.
- src/agent/tools/selfService.ts: forget_me, my_submissions, my_warnings,
  my_data.
- src/agent/tools/reportsMember.ts: report_content, withdraw_report,
  appeal_moderation + appealModerationCooldown/reserveAppealSlot +
  APPEAL_MODERATION_REASON_MAX_CHARS.
- src/agent/tools/feedback.ts: suggest_improvement, rate_answer,
  request_human_help + reserveHumanHelpRequestSlot +
  HUMAN_HELP_REQUEST_DAILY_LIMIT_PER_USER.
- src/agent/tools/prefs.ts: set_response_style, set_language_preference.
- src/agent/tools/reactions.ts: react_to_message + reactionDaily/
  reserveReactionDaily + ALLOWED_REACTION_EMOJI/REACTION_RATE_LIMIT_PER_DAY.
- src/agent/tools/social.ts: set_my_interests, who_is_into,
  set_helper_availability + find_helper (featureFlag cfg.findHelper.enabled),
  share_project, list_projects, request_project_connection +
  LIST_PROJECTS_DEFAULT_LIMIT.
- src/agent/tools/digestMember.ts: community_digest.
- src/agent/tools/projectNotes.ts: project_recall, project_note, project_list.
- PROJECT_NOTE_RETENTION_NOTICE moves to tools/helpers.ts — it is shared by
  forget_me (selfService.ts) and the still-unconverted purge_user_data
  closure, and shared state/constants live in helpers.ts by rule.
- tools.ts stays the barrel: every previously-exported symbol
  (EVENTS_LIST_LIMIT, CATCH_UP_*, APPEAL_MODERATION_REASON_MAX_CHARS,
  HUMAN_HELP_REQUEST_DAILY_LIMIT_PER_USER, ALLOWED_REACTION_EMOJI,
  REACTION_RATE_LIMIT_PER_DAY, LIST_PROJECTS_DEFAULT_LIMIT,
  PROJECT_NOTE_RETENTION_NOTICE) re-exported from its new home, so all
  existing import sites keep working. The server still registers
  [...registryTools, ...remaining closure tools] on one flat
  createSdkMcpServer (117 tools, unchanged — pinned by toolRegistry.test.ts).
- Three SECURITY source-scan tests (knowledge_search in tests/tools.test.ts,
  rate_answer in tests/unhelpfulAnswerEscalationRouter.test.ts,
  request_human_help in tests/humanHelpRequestRouter.test.ts) now read the
  handler's new domain file, with the handler-extraction regex tracking the
  defineTool wrapper shape — the never-calls-notifyAdmins assertions are
  unchanged, same treatment the stage 1 commit gave the feature_flags scans.
