+++ src/agent/tools.ts
@@ -4,62 +4,36 @@ import type { AdapterLookup, Platform, PlatformAdapter } from '../platforms/type
 import { assertAtLeast, atLeast, type CallerContext } from '../auth/rbac.js';
 import { normalizeMemberId } from '../auth/memberId.js';
 import { sanitizeName } from '../util/sanitizeName.js';
-import {
-  makeCalendarDayReserver,
-  makeCooldownReserver,
-  makeSlidingWindowReserver,
-} from '../util/rateReservation.js';
+import { makeCalendarDayReserver, makeSlidingWindowReserver } from '../util/rateReservation.js';
 import { isSuperAdmin, resolveRole } from '../auth/roles.js';
 import { config } from '../config.js';
 import { logger, hashId } from '../logger.js';
-import { WindowClosedError } from '../platforms/types.js';
-import { memoryHitJumpLink } from './discordLink.js';
 import {
   acceptKnowledgeCandidate,
   adminActivitySummary,
   addMemberNote,
-  answerFeedbackGrounding,
-  areKnowledgeEntriesLowRated,
-  candidateTopicAlreadyReviewed,
   clearAccessRequest,
   clearWarnings,
   countAccessRequests,
-  countActiveWarnings,
-  countMismatchedHelpfulRatings,
   countOpenAppeals,
   countOpenReports,
   countPendingKnowledgeCandidates,
   countPendingSuggestions,
-  countRecentDmReportsByReporterAndTarget,
-  countRepliesToUser,
-  createAnswerFeedback,
-  createContentReport,
-  createKnowledgeTip,
-  createModerationAppeal,
-  createSuggestion,
   clearUserSessions,
   declineKnowledgeCandidate,
   deleteKnowledge,
-  findKnowledgeCoveringTopic,
-  getActiveProjectById,
   getInteractionContentByMessageId,
   getKnowledgeContentById,
   deleteMemberNote,
   demoteAdmin,
   getMemberNote,
   getMemberRole,
-  getMyDataSummary,
-  hasConflictAmongIds,
   type KnowledgeCandidate,
   KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD,
-  KNOWLEDGE_TIP_CONTENT_MAX_CHARS,
-  KNOWLEDGE_TIP_RATE_LIMIT_PER_DAY,
-  KNOWLEDGE_TIP_TITLE_MAX_CHARS,
   type KnowledgeDuplicateMatch,
   listDuplicateKnowledge,
   listKnowledgeConflictCandidates,
   listKnowledgeCandidates,
-  listKnowledgeTopics,
   listBlockedUsers,
   listMemberNotes,
   listMemberWarnings,
@@ -68,7 +42,6 @@ import {
   isKnownConversation,
   isKnownMessage,
   isKnownUser,
-  isKnowledgeStale,
   isUserBlocked,
   linkMembers,
   listAccessRequests,
@@ -78,42 +51,25 @@ import {
   listContextDigests,
   listKnowledge,
   listKnowledgeFeedbackSummary,
-  listOwnAppeals,
-  listOwnKnowledgeCandidates,
-  listOwnProjectConnectionRequests,
-  listOwnProjects,
-  listOwnReports,
-  listOwnSuggestions,
-  listRecentInterests,
-  listRecentProjects,
   listReports,
   listRoster,
   listSuggestions,
-  MEMBER_PROJECT_CAP,
   mergeKnowledgeEntries,
   oldestAccessRequestAgeDays,
   oldestOpenAppealAgeDays,
   oldestOpenReportAgeDays,
   oldestPendingCandidateAgeDays,
   oldestPendingSuggestionAgeDays,
-  MEMBER_INTERESTS_MAX_CHARS,
   MODERATION_ACTION_KINDS,
   type ModerationAppeal,
   purgeUserData,
-  RATE_ANSWER_DAILY_LIMIT,
   recentAuditEntries,
-  recentConversationHistory,
   recentKnowledgeGapClusters,
-  findCrossedKnowledgeGapCluster,
   type CrossedKnowledgeGapCluster,
   recentModerationEntries,
   recentQuestionClusters,
   recentUnhelpfulFeedbackClusters,
-  recordKnowledgeGap,
-  recordKnowledgeRetrieval,
   removeMember,
-  removeMemberProject,
-  REPORT_RATE_LIMIT_PER_DAY,
   resolveContentReport,
   resolveDisplayName,
   resolveModerationAppeal,
@@ -123,36 +79,6 @@ import {
   resolveLinkedIdentities,
   saveKnowledge,
   getLanguagePreference,
-  setHelperAvailability,
-  setLanguagePreference,
-  setMemberInterests,
-  setResponseStyle,
-  searchMemberInterests,
-  searchMemberInterestsForSelf,
-  shareProject,
-  findHelperCandidates,
-  recordHelperNotificationIfUnderCap,
-  isFindHelperRequesterAtDailyCap,
-  FIND_HELPER_TOPIC_MAX_CHARS,
-  FIND_HELPER_REQUESTER_DAILY_LIMIT,
-  FIND_HELPER_WEEKLY_LIMIT_PER_HELPER,
-  withdrawOwnReports,
-  withdrawOwnKnowledgeTips,
-  SUGGESTION_MAX_CHARS,
-  SUGGESTION_RATE_LIMIT_PER_DAY,
-  PROJECT_CONNECTION_REQUESTER_DAILY_LIMIT,
-  PROJECT_DESCRIPTION_MAX_CHARS,
-  PROJECT_LINK_MAX_CHARS,
-  PROJECT_NAME_MAX_CHARS,
-  PROJECT_RATE_LIMIT_PER_DAY,
-  recordProjectConnectionIfUnderCap,
-  isProjectConnectionRequesterAtDailyCap,
-  searchKnowledge,
-  searchKnowledgeLexical,
-  searchProjectNotes,
-  saveProjectNote,
-  listVisibleProjects,
-  recordProjectNoteRetrieval,
   createProject,
   getProjectBySlug,
   addProjectMember,
@@ -161,68 +87,49 @@ import {
   unbindProjectSurface,
   archiveProject,
   unarchiveProject,
-  PROJECT_NOTE_CONTENT_MAX_CHARS,
-  PROJECT_NOTE_TITLE_MAX_CHARS,
-  PROJECT_NOTE_REFERENCE_URL_MAX_CHARS,
-  PROJECT_NOTE_RATE_LIMIT_PER_DAY,
   TEAM_PROJECT_NAME_MAX_CHARS,
   TEAM_PROJECT_BRIEF_MAX_CHARS,
   listAllProjects,
   listProjectMembers,
   listProjectSurfaces,
-  type KnowledgeSearchHit,
-  searchMemory,
-  searchProjects,
   unlinkMember,
   updateKnowledge,
   upsertMember,
   usageStats,
   engagementStats,
   userMessages,
-  WHO_IS_INTO_LIMIT,
 } from '../storage/repository.js';
-import { getCommunityGuidelines, getCommunityGuidelinesMi, updatePolicy } from '../storage/policies.js';
+import { updatePolicy } from '../storage/policies.js';
 import { recentChanges } from './changelog.js';
 import { redactSecrets } from './outbound.js';
 import { createIssue } from '../github/issues.js';
 import { triggerRedeploy } from './redeploy.js';
 import { formatNzEventTime } from '../util/nzTime.js';
 import { buildAdminDigestForAdmin } from '../adminDigest.js';
-import { buildMemberDigestContent } from '../memberDigest.js';
-import { formatStatusMessage, getStatusCache } from '../status/anthropicStatus.js';
 import { makeToolContext } from './tools/context.js';
 import { TOOL_REGISTRY } from './tools/index.js';
 import {
   formatAdminActivity,
   formatEngagementStats,
   formatFeatureFlags,
-  formatInterestResults,
-  formatKnowledgeSearchResults,
-  formatKnowledgeTopics,
   formatOtherConfiguredKnobs,
-  formatProjectResults,
-  formatRelativeAge,
   formatUsageStats,
   isoInstantSchema,
   parseIsoInstant,
   platformArg,
+  PROJECT_NOTE_RETENTION_NOTICE,
   resolveSanitizedLabel,
   text,
-  truncateForEcho,
   unreachableConversationRefusal,
   untrusted,
 } from './tools/helpers.js';
 import {
-  ackReportedMessage,
   applyManualWarnStrike,
   notifyAdminApproved,
-  notifyAppealFiled,
   notifyAppealResolved,
   notifyKnowledgeTipResolved,
   notifyMemberApproved,
-  notifyReportFiled,
   notifyReportResolved,
-  notifyReportWithdrawn,
   notifySuggestionResolved,
   notifyWarningsCleared,
 } from './tools/notify.js';
@@ -262,6 +169,7 @@ export {
   formatDevTeamJobStatus,
   formatDevTeamJobListEntry,
   formatDevTeamJobResult,
+  PROJECT_NOTE_RETENTION_NOTICE,
 } from './tools/helpers.js';
 export {
   notifyAdmins,
@@ -277,26 +185,12 @@ export {
   notifyWarningsCleared,
 } from './tools/notify.js';
 export { reserveDevTeamDispatchDaily } from './tools/devTeam.js';
-
-/** Per-message truncation shared by remember_search and catch_up (issue #167) so both quote the same amount of any one message. */
-const RECALL_TRUNCATION_CHARS = 400;
-
-/** catch_up (issue #167): default recap window when the caller doesn't ask for a specific one. */
-export const CATCH_UP_DEFAULT_HOURS = 24;
-
-/** catch_up: hard ceiling on the requested window, regardless of what `hours` asks for. */
-export const CATCH_UP_MAX_HOURS = 24 * 7;
-
-/**
- * catch_up's own row cap — deliberately NOT config.behaviour.memoryTopK
- * (tuned for a handful of embedding-similarity hits, not a whole-window
- * recap). Each row truncates to RECALL_TRUNCATION_CHARS (400) chars, so this
- * cap tops out around 40 * 400 = 16,000 chars (~4k tokens) of injected
- * untrusted context for one tool call — a bounded, deliberate slice of the
- * current turn's budget on top of the smaller automatic recall already
- * injected each turn.
- */
-export const CATCH_UP_MAX_MESSAGES = 40;
+export { EVENTS_LIST_LIMIT } from './tools/info.js';
+export { CATCH_UP_DEFAULT_HOURS, CATCH_UP_MAX_HOURS, CATCH_UP_MAX_MESSAGES } from './tools/memory.js';
+export { APPEAL_MODERATION_REASON_MAX_CHARS } from './tools/reportsMember.js';
+export { HUMAN_HELP_REQUEST_DAILY_LIMIT_PER_USER } from './tools/feedback.js';
+export { ALLOWED_REACTION_EMOJI, REACTION_RATE_LIMIT_PER_DAY } from './tools/reactions.js';
+export { LIST_PROJECTS_DEFAULT_LIMIT } from './tools/social.js';
 
 // Re-exported (not defined here — see the import above) so storage/
 // repository.ts's own `knowledgeCoversTopic` dedup guard (issue #102) can
@@ -304,9 +198,6 @@ export const CATCH_UP_MAX_MESSAGES = 40;
 // import cycle. See the full derivation comment on the definition in
 // repository.ts.
 export { KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD };
-
-/** list_projects' row cap for both the no-query (recent) and query (similarity) paths. */
-export const LIST_PROJECTS_DEFAULT_LIMIT = 8;
 
 /**
  * Cap on stored community guidelines text (issue #212). Bounded by Discord's
@@ -403,13 +294,6 @@ export const EVENT_DESCRIPTION_MAX_CHARS = 1000;
 export const EVENT_LOCATION_MAX_CHARS = 100;
 
 /**
- * appeal_moderation's optional free-text `reason` (issue #496) — same
- * bound treatment as `report_content`'s `reason`, since both are a short,
- * member-supplied explanation destined for an outbound admin DM.
- */
-export const APPEAL_MODERATION_REASON_MAX_CHARS = 500;
-
-/**
  * cancel_event's audit-only `reason` (issue #424) has no Discord field to
  * bound it against — same shape as report_content's `reason`, so the same
  * 500-char cap.
@@ -427,126 +311,12 @@ export const EVENT_CANCEL_REASON_MAX_CHARS = 500;
 const MEMBER_DM_FAILED_NOTE = " (Couldn't DM them the welcome message — they may not know yet.)";
 
 /**
- * Plain-language rundown of what a member can ask the bot to do, named by
- * behaviour rather than tool id (issue #92) — every entry in MEMBER_TOOLS
- * gets a line, most safety-relevant (report_content) first. Kept to a few
- * short lines deliberately: a wall of text reads worse than the terse blurb
- * it replaces.
- */
-const MEMBER_CAPABILITIES_TEXT =
-  'NZ Claude Community — a New Zealand group building with Claude and the Anthropic API. ' +
-  "Here's what you can ask me to do:\n" +
-  '- Flag harassment, spam, or a rule violation to admins ("report this"), or withdraw one filed by mistake\n' +
-  '- Ask admins to review a warning you think was a mistake ("appeal my warning")\n' +
-  '- Ask me for our community guidelines ("what are the rules here?")\n' +
-  '- Answer questions from curated community knowledge — just ask\n' +
-  '- Browse the topics our knowledge base covers, if you\'re not sure what to ask ("what do you know about?")\n' +
-  '- Search back through your own past messages for something said earlier\n' +
-  "- Check what I've stored about you, your active warnings, or your filed suggestions/reports\n" +
-  '- Catch you up on recent activity in this conversation ("what did I miss?")\n' +
-  '- Suggest how the bot or community could be better, or suggest a knowledge-base tip for other members ' +
-  'to find later, or withdraw one before an admin reviews it\n' +
-  '- Rate my last answer helpful or not\n' +
-  '- Ask to talk to a human community admin, if I\'m not getting you anywhere ("can I talk to a ' +
-  'human?")\n' +
-  '- Ask me to explain things more simply, or reply in te reo Māori ("keep it simple")\n' +
-  '- React to a message with an emoji instead of replying\n' +
-  '- Ask if a Claude/API problem is a known Anthropic outage, not your bug\n' +
-  '- Ask what meetups/events are coming up ("what\'s on?")\n' +
-  '- Share a project you\'ve built with the community, or browse what others have shared ("share my ' +
-  'project", "what has everyone built?")\n' +
-  "- Ask to connect with a project owner who's looking for collaborators (\"I'd like to help with that " +
-  'project")\n' +
-  '- Publish your own interests so other members can find you, or find members into a topic ("add me to ' +
-  'who\'s into RAG", "who\'s working on Discord bots?")\n' +
-  '- Ask if someone in the community can help with something you\'re stuck on ("can someone help with ' +
-  'X?"), or opt in/out of being notified for other members\' requests\n' +
-  '- Pull the community digest on demand\n' +
-  "- Record decisions in a project you're part of and search that project's shared memory later, or " +
-  'list your projects\n' +
-  '- Erase all your stored data any time ("forget me")';
-
-/**
- * Fixed-literal rundown of the WhatsApp `!`-prefixed text-command shortcuts
- * (issue #859), appended to MEMBER_CAPABILITIES_TEXT only for a WhatsApp
- * caller when `config.behaviour.whatsappTextCommandsEnabled` is true (issue
- * #872) — Discord already gets free discovery via its native `/` picker
- * (`SlashCommandBuilder.setDescription`, `src/platforms/discord/slashCommands.ts`),
- * which WhatsApp has no client-native equivalent of. No `!kb`: the existing
- * KNOWLEDGE_SHORTCUT_ENABLED shortcut already covers WhatsApp for that one
- * (#859's own decision). Never interpolates caller or message data — same
- * trust level as MEMBER_CAPABILITIES_TEXT.
- */
-const WHATSAPP_TEXT_COMMANDS_TEXT =
-  "You're on WhatsApp, so you can also use these zero-wait shortcuts:\n" +
-  '- `!whois <topic>` — find members into a topic\n' +
-  '- `!projects [query]` — browse the project showcase\n' +
-  '- `!guidelines` — community guidelines\n' +
-  "- `!digest` — this week's digest";
-
-/**
- * Plain-language rundown of what an admin can additionally ask the bot to
- * do, on top of MEMBER_CAPABILITIES_TEXT above (issue #367) — every entry in
- * ADMIN_TOOLS gets a mention, consolidated into behaviourally-related
- * bullets rather than 44 one-per-line entries, same discipline
- * MEMBER_CAPABILITIES_TEXT already uses (issue #311). Safety-relevant tools
- * (moderate, clear_warnings, archive_thread) come first, mirroring
- * MEMBER_CAPABILITIES_TEXT's own "most safety-relevant first" convention.
- * No interpolation of any runtime/tool argument — static text only, same
- * trust level as MEMBER_CAPABILITIES_TEXT.
- */
-const ADMIN_CAPABILITIES_TEXT =
-  'As an admin, you also have:\n' +
-  "- Moderate the community: warn, mute, kick, or remove a message, clear a member's warnings, archive a Discord thread, review the moderation history log, pull one member's full warning history, list everyone who's currently muted, list who's currently blocked on WhatsApp, or review and resolve filed appeals\n" +
-  "- Manage membership: add a new member, remove a member, link a member's cross-platform identity, or unlink a member's cross-platform identity\n" +
-  '- Review flagged content reports and resolve each report, review suggestions members submit and resolve each suggestion, see how members rated my answers, check which knowledge entries are rated poorly, and review recurring unhelpful-answer themes across all answers\n' +
-  '- Post to the community: make an announcement, create a poll or end one poll early, open a Discord thread, or schedule/cancel an event\n' +
-  '- Curate the knowledge base: save a new knowledge entry, browse knowledge entries, edit a knowledge entry, delete a knowledge entry, or merge two entries together, and check for near-duplicate entries or conflicting entries\n' +
-  "- Review knowledge candidates, accept a candidate or decline a candidate, track knowledge gaps (questions I couldn't answer), recurring question clusters, raw context digests, pull your own admin-digest snapshot on demand, get a review-queue roll-up of all five review queues at once, or check how quickly I've been answering members (response latency)\n" +
-  '- See who is waiting for access, or who has joined or left the server\n' +
-  "- Add a note about a member, review notes on a member, delete a note, or look up a member's history across conversations\n" +
-  '- Set the community guidelines or the welcome message shown to new members\n' +
-  '- Assign a Discord role, remove a Discord role, or list which roles are available to assign\n' +
-  "- Set up team projects: create one, give a member access, take a member's access away, allow or " +
-  'stop it being discussed here, review who has access, or archive a finished project and bring it ' +
-  'back again\n' +
-  '- Generate an image, or check recent changes to the bot and community (the changelog)';
-
-/**
- * Plain-language rundown of what a super admin can additionally ask the bot
- * to do, on top of MEMBER_CAPABILITIES_TEXT and ADMIN_CAPABILITIES_TEXT above
- * (issue #582) — every entry in SUPER_ADMIN_TOOLS gets a mention,
- * consolidated into behaviourally-related bullets rather than 19 one-per-line
- * entries, same discipline ADMIN_CAPABILITIES_TEXT already uses (issue #367).
- * No interpolation of any runtime/tool argument — static text only, same
- * trust level as its two siblings.
- */
-const SUPER_ADMIN_CAPABILITIES_TEXT =
-  'As a super admin, you also have:\n' +
-  '- Grant or revoke admin status for a member\n' +
-  '- Pause or resume the bot, view audit logs, review admin activity, list current admins, ' +
-  'or check usage/engagement stats\n' +
-  '- Erase all of a user\'s stored data on request ("purge their data")\n' +
-  '- Change bot-wide policy settings, or trigger a redeploy of the bot\n' +
-  '- See which optional feature flags are currently on or off\n' +
-  '- File a GitHub issue suggesting an improvement\n' +
-  '- Dispatch a remote dev-team job to assess or deliver a change, check its status, fetch its result, ' +
-  "turn a completed assessment into a tracked backlog, list an assessment's findings, or re-check one finding";
-
-/**
  * Fixed, static note appended to `grant_admin`'s reply when
  * `notifyAdminApproved` reports the promotion DM did not land (issue #556) —
  * mirrors `MEMBER_DM_FAILED_NOTE`'s rationale exactly, with its own wording
  * since this is a promotion, not a fresh membership.
  */
 const ADMIN_DM_FAILED_NOTE = " (Couldn't DM them about the promotion — they may not know yet.)";
-
-/**
- * Fixed cap on how many upcoming events `list_events` returns (issue #388) —
- * a small hardcoded constant over a config knob, matching this repo's
- * existing convention for tool-shape limits (e.g. `GATED_NOTICE_MAX_ADMIN_NAMES`).
- */
-export const EVENTS_LIST_LIMIT = 10;
 
 /**
  * After a role change (grant_admin/revoke_admin) commits, reset the target's
@@ -654,77 +424,12 @@ export const reserveVoiceTranscriptionSlot = makeSlidingWindowReserver(60 * 60 *
 const reserveWarnSlot = makeSlidingWindowReserver(60 * 60 * 1000);
 
 /**
- * request_human_help timestamps per caller (`platform:userId`), for its own
- * rolling-24h daily cap (HUMAN_HELP_REQUEST_DAILY_LIMIT_PER_USER, issue
- * #808) — in-memory, same sliding-window shape as `reservePollSlot`/
- * `reserveAnnounceSlot` above, not a new table. This bounds a single
- * caller's own worst-case share of the shared guild-wide
- * ESCALATION_RATE_LIMIT_PER_HOUR budget: without it one member spamming the
- * tool could alone exhaust that hourly cap and starve every other member's
- * max-turns/thumbs-down escalations for the rest of the hour.
- */
-export const HUMAN_HELP_REQUEST_DAILY_LIMIT_PER_USER = 3;
-
-/**
- * Reserve one request_human_help slot for `key` against a rolling 24h cap —
- * same sliding-window shape as `reservePollSlot`, but a day-long window
- * (this bounds a caller's daily share, not a per-conversation hourly burst).
- * Returns false without reserving if `key` already hit `limit` within the
- * last 24h.
- */
-const reserveHumanHelpRequestSlot = makeSlidingWindowReserver(24 * 60 * 60 * 1000);
-
-/**
  * Reserve one announce slot for `conversationId` against a rolling hourly
  * cap (ANNOUNCE_RATE_LIMIT_PER_HOUR), same sliding-window shape as
  * `reservePollSlot`. Returns false without reserving if the conversation
  * already hit `limit` within the last hour.
  */
 const reserveAnnounceSlot = makeSlidingWindowReserver(60 * 60 * 1000);
-
-/**
- * appeal_moderation last-fired timestamp per CALLER (`platform:userId`), for
- * its per-caller cooldown (`MODERATION_APPEAL_COOLDOWN_HOURS`, issue #496).
- * Scoped to the caller rather than the conversation — unlike every
- * `reserve*Slot` cap above — since an appeal is inherently about one
- * person's own status. In-memory/best-effort for the MVP (no new table): a
- * restart merely permits one extra appeal DM, harmless for a non-destructive
- * notification.
- */
-const appealModerationCooldown = makeCooldownReserver();
-
-/**
- * Reserve one appeal_moderation slot for `key` against a rolling per-caller
- * cooldown. Returns false without reserving if `key` already appealed within
- * `cooldownHours`.
- */
-function reserveAppealSlot(key: string, cooldownHours: number): boolean {
-  return appealModerationCooldown(key, cooldownHours * 60 * 60 * 1000);
-}
-
-/**
- * Closed emoji allowlist for `react_to_message` (issue #231) — positive/
- * neutral only, deliberately excluding anything that could read as the bot
- * editorialising against a member (no 👎). Never interpolate a model-supplied
- * emoji string into the Discord API; only one of these fixed values ever
- * reaches `adapter.reactToMessage`, matching the closed-enum discipline
- * `set_language_preference` already uses for untrusted-string inputs.
- */
-export const ALLOWED_REACTION_EMOJI = ['✅', '👍', '👀', '🎉'] as const;
-
-/** Per-user reaction tally for the current UTC day (anti-spam on the bot's own identity; issue #231). */
-export const REACTION_RATE_LIMIT_PER_DAY = 20;
-const reactionDaily = makeCalendarDayReserver();
-
-/**
- * Reserve one reaction slot for `key` against today's per-user cap, same
- * restart-resets-the-window shape as `reserveImageGenDaily` — acceptable here
- * because a reaction is far lower-consequence than an image-gen subprocess
- * spawn, so an in-memory (not DB) cap is proportionate and needs no migration.
- */
-function reserveReactionDaily(key: string): boolean {
-  return reactionDaily(key, REACTION_RATE_LIMIT_PER_DAY);
-}
 
 /**
  * Turn-scoped, mutable correlation state threaded in from `execTurn` (issue
@@ -783,24 +488,6 @@ export interface ToolServerTurnState {
 }
 
 /**
- * Shared by `forget_me` and `purge_user_data`'s CONFIRM prompt and
- * post-confirm reply (issue #930), so the two can't drift the way the
- * original "delete ALL of X's stored data" wording did once #929 made
- * project-note retention (docs/SECURITY.md §25) the one partial-erasure
- * exception in this codebase. States all three facts an erasure claim has to
- * convey when the target is a project member: membership is deleted and
- * access ends immediately on every platform; project notes they authored are
- * KEPT with the authorship link removed; and removing that link does not
- * scrub personal information the note's own text may contain, so a note that
- * names them still names them after erasure. Deliberately generic — never
- * enumerates project names — since the CONFIRM prompt fires at the exact
- * moment someone is asserting a privacy right and reciting project names into
- * a possibly-public conversation would itself be a small exposure.
- */
-export const PROJECT_NOTE_RETENTION_NOTICE =
-  "if they're in a project, membership is deleted immediately on every platform, but any project notes they wrote are kept with the authorship link removed — that removes the link only, not personal information the note's own text may contain, so a note naming them still names them";
-
-/**
  * Build the in-process MCP tool server for one agent turn. The tools close
  * over the caller context and the adapter handling this conversation, so
  * RBAC and platform routing are baked in. Layers:
@@ -828,1735 +515,6 @@ export function buildToolServer(
     tool(def.name, def.description, def.schema, (args) => def.handler(args, ctx), {
       annotations: { readOnlyHint: def.readOnlyHint },
     }),
-  );
-
-  // --- Member tools ----------------------------------------------------------
-
-  const communityInfo = tool(
-    'community_info',
-    'Tell the caller, in concrete terms, what they can ask this bot to do. Call this whenever someone ' +
-      'asks "what can you do?", "how do I report someone?", or otherwise wants a capability rundown — ' +
-      'do not answer that from general knowledge alone.',
-    {},
-    async () => {
-      const memberSegment =
-        caller.platform === 'whatsapp' &&
-        config.behaviour.whatsappTextCommandsEnabled &&
-        atLeast(caller.role, 'member')
-          ? `${MEMBER_CAPABILITIES_TEXT}\n${WHATSAPP_TEXT_COMMANDS_TEXT}`
-          : MEMBER_CAPABILITIES_TEXT;
-      if (caller.role === 'super_admin') {
-        return text(`${memberSegment}\n${ADMIN_CAPABILITIES_TEXT}\n${SUPER_ADMIN_CAPABILITIES_TEXT}`);
-      }
-      if (caller.role === 'admin') {
-        return text(`${memberSegment}\n${ADMIN_CAPABILITIES_TEXT}`);
-      }
-      return text(memberSegment);
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const communityGuidelines = tool(
-    'community_guidelines',
-    "Return this community's guidelines/rules, exactly as an admin set them. Call this whenever someone " +
-      'asks "what are the rules?", "what am I not allowed to do?", or wants to know why they were warned ' +
-      'or muted. Relay the returned text to the caller verbatim — do not summarise, paraphrase, or add to it.',
-    {},
-    async () => {
-      const languagePreference = await getLanguagePreference(caller.platform, caller.userId);
-      const guidelines =
-        languagePreference === 'mi'
-          ? ((await getCommunityGuidelinesMi()) ?? (await getCommunityGuidelines()))
-          : await getCommunityGuidelines();
-      return text(guidelines ?? 'No community guidelines have been set yet — ask an admin.');
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const checkStatus = tool(
-    'check_status',
-    'Check whether Anthropic has a known service incident right now — call this when a member reports an ' +
-      "error, timeout, or unexpected behaviour from Claude/the API and wants to know if it's a known Anthropic " +
-      "problem rather than something on their end. Read-only, no arguments, sourced from Anthropic's own public " +
-      'status page (a background poll, never a live fetch on this call).',
-    {},
-    async () => text(formatStatusMessage(getStatusCache(), Date.now())),
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const listEvents = tool(
-    'list_events',
-    'List upcoming Discord scheduled meetups/events (id, name, start/end time, location) — call this when ' +
-      'someone asks "what\'s coming up?", "when\'s the next meetup?", or similar, instead of guessing from ' +
-      'general knowledge or stale knowledge-base entries. Also the only way to discover a valid eventId for ' +
-      'cancel_event. Read-only, no arguments, sourced live from ' +
-      "Discord's own Scheduled Events (the read counterpart to create_event). Discord-only.",
-    {},
-    async () => {
-      if (!adapter.listUpcomingEvents) {
-        return text(`Event listings aren't available on ${caller.platform}.`, true);
-      }
-      const events = await adapter.listUpcomingEvents(EVENTS_LIST_LIMIT);
-      if (events.length === 0) return text('No upcoming events.');
-      return text(
-        events
-          .map((e) => {
-            const when = e.scheduledEndAt
-              ? `${formatNzEventTime(e.scheduledStartAt)} – ${formatNzEventTime(e.scheduledEndAt)}`
-              : formatNzEventTime(e.scheduledStartAt);
-            const desc = e.description ? `: ${e.description}` : '';
-            return `- ${e.name} (${when}) @ ${e.location}${desc} [id: ${e.id}]`;
-          })
-          .join('\n'),
-      );
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const knowledgeSearch = tool(
-    'knowledge_search',
-    'Search curated community knowledge (FAQs, rules, resources admins have saved).',
-    { query: z.string().describe('Topic to look up') },
-    async (args) => {
-      const hits = await searchKnowledge(args.query, {
-        platform: caller.platform,
-        conversationId: caller.conversationId,
-      });
-      // Fire-and-forget usage tracking (issue #134) — entries that clear the
-      // relevance floor are "used"; ones that exist but fall below it are
-      // not. Never awaited and errors are swallowed here (not inside
-      // recordKnowledgeRetrieval) so a counter-write failure can never delay
-      // or fail this member-facing search, mirroring notifySuperAdmins'
-      // inline-catch, non-awaited style.
-      const relevantIds = hits
-        .filter((h) => h.similarity >= KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD)
-        .map((h) => h.id);
-      recordKnowledgeRetrieval(relevantIds).catch((err) =>
-        logger.warn({ err }, 'Knowledge retrieval count update failed'),
-      );
-      // Best-effort knowledge_search-hit correlation (issue #411): `hits` is
-      // already ordered by similarity descending (searchKnowledge's `ORDER
-      // BY embedding <=> $1`), so relevantIds[0] — if any cleared the floor —
-      // is this call's top-scoring hit. Only overwrite on a QUALIFYING call;
-      // a later call in the same turn whose hits all miss the floor must
-      // never clobber an earlier qualifying id with null (acceptance
-      // criterion #3: last *qualifying* call wins, not last call).
-      if (turnState && relevantIds.length > 0) {
-        turnState.lastKnowledgeHitId = relevantIds[0];
-      }
-      // Live conflict-candidate check (issue #389): only the ids that
-      // cleared the relevance floor for THIS query, restricted to a
-      // scoped, LIMIT-1 self-join — never the full-table audit
-      // `listKnowledgeConflictCandidates` runs. Skipped entirely below 2
-      // ids, matching hasConflictAmongIds' own zero-query short-circuit.
-      // Fail-safe like the low-rated caveat below: the conflict note is a
-      // purely advisory badge, so a lookup failure must degrade to "no note"
-      // rather than discarding the hits we already fetched successfully.
-      const hasConflict =
-        relevantIds.length >= 2
-          ? await hasConflictAmongIds(relevantIds).catch((err) => {
-              logger.warn({ err }, 'Knowledge conflict check failed; omitting the conflict note');
-              return false;
-            })
-          : false;
-      // Member-facing low-rated-answer caveat (issue #432) — the display-side
-      // counterpart to #337's shortcut-only caveat: this is the dominant
-      // answer path (below the shortcut's 0.9-cosine ceiling), so gating and
-      // fail-safe behaviour mirror sendKnowledgeShortcut's own exactly. The
-      // extra query only runs when the feature is enabled AND at least one
-      // hit cleared the relevance floor, matching hasConflictAmongIds' own
-      // zero-query short-circuit for a too-small input.
-      const lowRatedIds =
-        config.behaviour.knowledgeLowRatedCaveatMinUnhelpful > 0 && relevantIds.length > 0
-          ? await areKnowledgeEntriesLowRated(
-              relevantIds,
-              config.behaviour.knowledgeLowRatedCaveatMinUnhelpful,
-            ).catch((err) => {
-              logger.warn({ err }, 'Knowledge low-rated caveat lookup failed; omitting the caveat');
-              return new Set<number>();
-            })
-          : new Set<number>();
-      // Lexical fallback (issue #362): only on the below-floor-miss branch
-      // below — semantic search had candidates but NONE cleared the
-      // relevance floor. Dense sentence embeddings underweight rare,
-      // SNAKE_CASE/camelCase identifiers and error codes, so a query that's
-      // literally a string inside an entry can still miss; try a
-      // substring-robust trigram match before accepting this as a gap. When
-      // semantic search already found a relevant hit, this never runs —
-      // output is byte-identical to before issue #362 for the common case.
-      let lexicalHits: Awaited<ReturnType<typeof searchKnowledgeLexical>> = [];
-      if (hits.length > 0 && relevantIds.length === 0) {
-        // Fail-safe, same reasoning as the conflict/low-rated lookups above:
-        // this is a supplementary second attempt on the below-floor branch,
-        // so a failure here must degrade to "no lexical hits" and still show
-        // the semantic results, never replace them with a raw DB error.
-        lexicalHits = await searchKnowledgeLexical(args.query, {
-          platform: caller.platform,
-          conversationId: caller.conversationId,
-        }).catch((err) => {
-          logger.warn({ err }, 'Knowledge lexical fallback failed; returning semantic results only');
-          return [];
-        });
-      }
-      if (lexicalHits.length > 0) {
-        recordKnowledgeRetrieval(lexicalHits.map((h) => h.id)).catch((err) =>
-          logger.warn({ err }, 'Knowledge retrieval count update failed'),
-        );
-      } else if (hits.length > 0 && relevantIds.length === 0) {
-        // Below-floor miss tracking (issue #208): only when hits existed but
-        // NONE cleared the floor (semantic or, now, lexical) — never on a
-        // plain empty result set, which is indistinguishable from a
-        // searchKnowledge embed() failure and would otherwise log every
-        // outage query as a false "gap". Fire-and-forget, same non-blocking
-        // style as the retrieval-count bump above — UNLESS the real-time
-        // cluster alert (issue #650) is enabled, in which case the insert is
-        // awaited (one extra fast DB round-trip) so the immediately-following
-        // cluster-threshold check has the new row's id to key off. With the
-        // flag off (the default) this branch is byte-identical to before
-        // #650 — no extra query, no await, no DM (acceptance criterion 3).
-        if (config.knowledgeGapAlert.enabled && turnState) {
-          const gapResult = await recordKnowledgeGap(
-            caller.platform,
-            caller.conversationId,
-            caller.userId,
-            args.query,
-          ).catch((err) => {
-            logger.warn({ err }, 'Knowledge gap recording failed');
-            return null;
-          });
-          if (gapResult && gapResult !== 'rate_limited') {
-            const crossed = await findCrossedKnowledgeGapCluster(
-              caller.conversationId,
-              gapResult.id,
-              config.knowledgeGapAlert.threshold,
-            ).catch((err) => {
-              logger.warn({ err }, 'Knowledge gap cluster threshold check failed');
-              return null;
-            });
-            if (crossed) turnState.knowledgeGapCluster = crossed;
-          }
-        } else {
-          recordKnowledgeGap(caller.platform, caller.conversationId, caller.userId, args.query).catch((err) =>
-            logger.warn({ err }, 'Knowledge gap recording failed'),
-          );
-        }
-      }
-      const finalHits: Array<KnowledgeSearchHit & { viaLexical?: boolean }> =
-        lexicalHits.length > 0 ? [...hits, ...lexicalHits.map((h) => ({ ...h, viaLexical: true }))] : hits;
-      // Real-time stale-knowledge admin nudge (issue #701): computed over
-      // exactly the hits `formatKnowledgeSearchResults` below will actually
-      // render (its own identical `viaLexical || similarity >= floor`
-      // filter) — never a hit that exists but isn't shown. Gated on the flag
-      // FIRST so this is a no-op (no isKnowledgeStale call, no turnState
-      // write) when off, matching acceptance criterion 4's byte-identical
-      // default. `notifyAdmins` itself is never called from this file — see
-      // its own doc comment; router.ts does the gate+stamp+notify post-turn.
-      if (config.knowledgeStaleAlert.enabled && turnState) {
-        for (const h of finalHits) {
-          if (!h.viaLexical && h.similarity < KNOWLEDGE_SEARCH_RELEVANCE_THRESHOLD) continue;
-          if (
-            isKnowledgeStale(
-              { updatedAt: h.updatedAt, lastRetrievedAt: h.lastRetrievedAt ?? null },
-              config.adminDigest.knowledgeStaleDays,
-              config.adminDigest.knowledgeStaleMaxAgeDays,
-            )
-          ) {
-            (turnState.staleKnowledgeAlertIds ??= []).push(h.id);
-          }
-        }
-      }
-      return text(
-        formatKnowledgeSearchResults(
-          finalHits,
-          config.adminDigest.knowledgeStaleDays,
-          config.adminDigest.knowledgeStaleMaxAgeDays,
-          hasConflict,
-          lowRatedIds,
-        ),
-      );
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const listKnowledgeTopicsTool = tool(
-    'list_knowledge_topics',
-    'Browse the titles of what the community knowledge base covers — the proactive counterpart to ' +
-      "knowledge_search for a member who doesn't yet know the right words to search for. Titles only, " +
-      'no arguments, no content — call knowledge_search for an actual answer once you know what to ask.',
-    {},
-    async () => {
-      const { titles, totalCount } = await listKnowledgeTopics(
-        { platform: caller.platform, conversationId: caller.conversationId },
-        config.behaviour.knowledgeTopicsListLimit,
-      );
-      return text(formatKnowledgeTopics(titles, totalCount));
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const rememberSearch = tool(
-    'remember_search',
-    'Search past interactions for relevant context. Members search the current conversation; admins may search conversations they are in; super admins may search everything.',
-    {
-      query: z.string().describe('What to search for in past conversations'),
-      scope: z
-        .enum(['conversation', 'mine', 'all'])
-        .optional()
-        .describe(
-          "'conversation' (default) = this conversation; 'mine' (admin) = all conversations you are in; 'all' (super admin) = every conversation on both platforms",
-        ),
-    },
-    async (args) => {
-      const scope = args.scope ?? 'conversation';
-      let hits;
-      if (scope === 'all') {
-        assertAtLeast(caller.role, 'super_admin', 'remember_search:all');
-        hits = await searchMemory(args.query, {});
-      } else if (scope === 'mine') {
-        assertAtLeast(caller.role, 'admin', 'remember_search:mine');
-        const allowed = await callerScope();
-        hits = await searchMemory(args.query, {
-          platform: caller.platform,
-          ...(allowed ? { conversationIds: allowed } : {}),
-        });
-      } else {
-        hits = await searchMemory(args.query, {
-          platform: caller.platform,
-          conversationId: caller.conversationId,
-        });
-      }
-      if (hits.length === 0) return text('No relevant past interactions found.');
-      return text(
-        untrusted(
-          'Search results',
-          hits
-            .map((h, i) => {
-              const link = memoryHitJumpLink(h, config.discord.guildId);
-              // Sanitize the recalled author name (untrusted platform display
-              // name): untrusted() strips angle brackets but not newlines, so a
-              // `\n\n[SYSTEM] ...` nickname would otherwise land as an apparent
-              // standalone directive inside this result (finding A).
-              const name = sanitizeName(h.userName);
-              return `${i + 1}. (${(h.similarity * 100).toFixed(0)}% match) [${h.direction}${name ? ` by ${name}` : ''}] ${h.content.slice(0, RECALL_TRUNCATION_CHARS)}${link ? ` (${link})` : ''}`;
-            })
-            .join('\n'),
-        ),
-      );
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const forgetMe = tool(
-    'forget_me',
-    "Delete the requester's own stored data from the bot's memory (privacy request): their messages, " +
-      'plus any knowledge entries, content reports, suggestions, roster entry, and admin notes tied to ' +
-      'them — across linked identities. Requires confirmation.',
-    {},
-    async () =>
-      requireConfirm(
-        `erase ${caller.userName}'s stored data on ${caller.platform} — messages, and any knowledge entries, content reports, suggestions, roster entry, or admin notes tied to them, across linked identities; ${PROJECT_NOTE_RETENTION_NOTICE}`,
-        // Self-scoped: whatever tier the caller is, they can only ever purge
-        // their OWN data. An open-mode guest (whose content IS stored) can
-        // reach this tool, so gating the confirm at 'member' made their
-        // CONFIRM fail the tier re-check and report a false "your permissions
-        // changed". 'guest' is the correct floor for a self-scoped purge.
-        'guest',
-        async () => {
-          const n = await purgeUserData(caller.platform, caller.userId);
-          return `Deleted ${n} stored record(s) for ${caller.userName}; ${PROJECT_NOTE_RETENTION_NOTICE}.`;
-        },
-      ),
-  );
-
-  const reportContent = tool(
-    'report_content',
-    'Report harassment, spam, or a rule violation in this conversation to its admins for review. ' +
-      'Only confirms the report was recorded — it does not take any moderation action itself.',
-    {
-      reason: z.string().min(1).max(500).describe('What happened, in your own words (max 500 characters)'),
-      targetUserId: z.string().optional().describe('Platform user id of the person being reported, if known'),
-      messageId: z.string().optional().describe('The specific message id being reported, if known'),
-    },
-    async (args) => {
-      // targetUserId is reporter-supplied and unauthenticated — unlike
-      // moderate/clear_warnings (admin-only, already gated by isKnownUser),
-      // any member can name anyone here. Since target_user_id also drives the
-      // accused-admin visibility exclusion (listReports/countOpenReports/
-      // resolveContentReport), an unverified id could be used to blind an
-      // unrelated admin from a report that isn't about them at all. Only a
-      // target the bot has actually seen before is trusted to drive that
-      // exclusion; an unknown/typo'd id is dropped rather than stored
-      // (issue #197 review).
-      const targetUserId =
-        args.targetUserId && (await isKnownUser(caller.platform, args.targetUserId))
-          ? args.targetUserId
-          : undefined;
-      const created = await createContentReport({
-        platform: caller.platform,
-        reporterUserId: caller.userId,
-        reporterName: caller.userName,
-        conversationId: caller.conversationId,
-        targetUserId,
-        messageId: args.messageId,
-        reason: args.reason,
-        isDirect: caller.isDirect,
-      });
-      if (!created) {
-        return text(
-          `You've already submitted ${REPORT_RATE_LIMIT_PER_DAY} reports in the last 24 hours. ` +
-            'Please wait before submitting another, or contact an admin directly if this is urgent.',
-          true,
-        );
-      }
-      // Only computed for a DM report naming a known target — exactly the
-      // case the accused-admin exclusion applies to (issue #305). Inclusive
-      // of the just-inserted row, so this count reaching the threshold on
-      // the report that crosses it is what triggers the alert line.
-      const recentSameTargetCount =
-        caller.isDirect && targetUserId
-          ? await countRecentDmReportsByReporterAndTarget(caller.platform, caller.userId, targetUserId)
-          : undefined;
-      void notifyReportFiled(adapterFor, {
-        id: created.id,
-        reporterUserId: caller.userId,
-        reporterName: caller.userName,
-        conversationId: caller.conversationId,
-        targetUserId,
-        messageId: args.messageId,
-        reason: args.reason,
-        recentSameTargetCount,
-      });
-      ackReportedMessage(adapter, caller.platform, caller.conversationId, args.messageId);
-      return text(`Report #${created.id} recorded for this conversation's admins. Thanks for flagging it.`);
-    },
-  );
-
-  const withdrawReport = tool(
-    'withdraw_report',
-    'Withdraw your OWN previously-filed content report(s) — use this if you filed one by mistake or as a ' +
-      'joke and no longer want it reviewed. It only ever affects reports YOU filed; it cannot touch anyone ' +
-      "else's. The report is marked withdrawn and kept on record (not deleted), and the admins are notified.",
-    {},
-    async () => {
-      const ids = await withdrawOwnReports(caller.platform, caller.userId);
-      if (ids.length === 0) {
-        return text('You have no open reports to withdraw.', true);
-      }
-      void notifyReportWithdrawn(adapterFor, {
-        ids,
-        reporterUserId: caller.userId,
-        reporterName: caller.userName,
-      });
-      const list = ids.map((id) => `#${id}`).join(', ');
-      return text(
-        `Withdrew your report${ids.length > 1 ? 's' : ''} ${list}. ` +
-          "They won't be actioned; the admins have been notified of the withdrawal.",
-      );
-    },
-    { annotations: { readOnlyHint: false } },
-  );
-
-  const withdrawKnowledgeTip = tool(
-    'withdraw_knowledge_tip',
-    'Withdraw your OWN still-pending suggest_knowledge tip(s) — use this if you filed one by mistake, as a ' +
-      'joke, or want to fix it before an admin reviews it. It only ever affects tips YOU filed and only ones ' +
-      "still pending; it cannot touch anyone else's tip, a machine-drafted candidate, or a tip already " +
-      'reviewed. The tip is marked withdrawn and kept on record (not deleted).',
-    {},
-    async () => {
-      const ids = await withdrawOwnKnowledgeTips(caller.platform, caller.userId);
-      if (ids.length === 0) {
-        return text('You have no pending knowledge tips to withdraw.', true);
-      }
-      const list = ids.map((id) => `#${id}`).join(', ');
-      return text(
-        `Withdrew your knowledge tip${ids.length > 1 ? 's' : ''} ${list}. ` +
-          "They won't be reviewed — feel free to resubmit a better version with suggest_knowledge.",
-      );
-    },
-    { annotations: { readOnlyHint: false } },
-  );
-
-  const mySubmissions = tool(
-    'my_submissions',
-    "List the caller's OWN previously-filed suggestions, content reports, moderation appeals, knowledge " +
-      'tips, and sent project-connection requests — id, a short content preview, current status, and when ' +
-      'each was filed. Use this when a member asks what happened to something they submitted earlier (e.g. ' +
-      '"what happened to my report?"). The connection-requests section is a plain RECEIPT (what you asked, ' +
-      "when) — request_project_connection has no accept/decline state, so there's no status to show, and a " +
-      'capped/refused attempt is never recorded so it never appears here either. Never returns another ' +
-      "member's content or the reviewing admin's identity — only the shared admin queue " +
-      '(list_suggestions/list_reports/list_appeals/list_knowledge_candidates) exposes that, and this tool ' +
-      'never reaches it.',
-    {},
-    async () => {
-      const [suggestions, reports, appeals, knowledgeTips, connectionRequests] = await Promise.all([
-        listOwnSuggestions(caller.platform, caller.userId, 10),
-        listOwnReports(caller.platform, caller.userId, 10),
-        listOwnAppeals(caller.platform, caller.userId, 10),
-        listOwnKnowledgeCandidates(caller.platform, caller.userId, 10),
-        listOwnProjectConnectionRequests(caller.platform, caller.userId, 10),
-      ]);
-
-      if (
-        suggestions.length === 0 &&
-        reports.length === 0 &&
-        appeals.length === 0 &&
-        knowledgeTips.length === 0 &&
-        connectionRequests.length === 0
-      ) {
-        return text("You haven't filed any suggestions or reports yet.", true);
-      }
-
-      const lines: string[] = [];
-      if (suggestions.length > 0) {
-        lines.push('Your suggestions:');
-        for (const s of suggestions) {
-          lines.push(
-            `- #${s.id} [${s.status}] ${truncateForEcho(s.content)} — filed ${formatRelativeAge(s.createdAt)}`,
-          );
-        }
-      }
-      if (reports.length > 0) {
-        if (lines.length > 0) lines.push('');
-        lines.push('Your reports:');
-        for (const r of reports) {
-          lines.push(
-            `- #${r.id} [${r.status}] ${truncateForEcho(r.reason)} — filed ${formatRelativeAge(r.createdAt)}`,
-          );
-        }
-      }
-      if (appeals.length > 0) {
-        if (lines.length > 0) lines.push('');
-        lines.push('Your appeals:');
-        for (const a of appeals) {
-          const reason = a.reason ? truncateForEcho(a.reason) : 'no reason given';
-          lines.push(`- #${a.id} [${a.status}] ${reason} — filed ${formatRelativeAge(a.createdAt)}`);
-        }
-      }
-      if (knowledgeTips.length > 0) {
-        if (lines.length > 0) lines.push('');
-        lines.push('Your knowledge tips:');
-        for (const k of knowledgeTips) {
-          // "used N times" only for an accepted tip with a positive retrieval
-          // count (issue #880) — never "used 0 times" for an accepted-but-
-          // unretrieved or non-accepted tip, which would read as discouraging.
-          const impact =
-            k.status === 'accepted' && k.retrievalCount && k.retrievalCount > 0
-              ? ` — used ${k.retrievalCount} time${k.retrievalCount === 1 ? '' : 's'} in answers so far`
-              : '';
-          lines.push(
-            `- #${k.id} [${k.status}] ${truncateForEcho(k.title)} — filed ${formatRelativeAge(k.createdAt)}${impact}`,
-          );
-        }
-      }
-      if (connectionRequests.length > 0) {
-        if (lines.length > 0) lines.push('');
-        lines.push('Your connection requests:');
-        for (const c of connectionRequests) {
-          // No status column exists (issue #908) — this is a receipt, not a
-          // tracker. A since-removed/purged project reads back null; say so
-          // rather than rendering a blank or throwing.
-          const projectLabel = c.projectName ?? 'a project that is no longer listed';
-          lines.push(`- #${c.id} — ${projectLabel} — filed ${formatRelativeAge(c.createdAt)}`);
-        }
-      }
-      return text(lines.join('\n'));
-    },
-  );
-
-  const myWarnings = tool(
-    'my_warnings',
-    "Check the caller's OWN active auto-moderation warning count and the configured limit — use this when " +
-      'a member asks how many warnings they have or whether they can still post. Always scoped to the ' +
-      "caller's own platform/user id, never a model-supplied identifier. Never includes a warning's reason " +
-      'or excerpt — that context stays admin-only (see list_member_warnings).',
-    {},
-    async () => {
-      const limit = config.moderation.strikeLimit;
-      const windowDays = config.moderation.strikeWindowDays;
-      // Report on the UNWINDOWED count. A mute is only ever lifted by
-      // clear_warnings, never by strikes aging out of the window, so a member
-      // whose strikes have aged out of the window can still be blocked;
-      // reporting the windowed count alone told them "you have no active
-      // warnings" while they were still at/over the limit (advisory F5). This
-      // deliberately does NOT claim a live Discord mute — the tool can't read
-      // the role state (issue #182) — only the caller's count vs. the limit.
-      // When no window is configured the two counts are identical, so the
-      // extra read is skipped.
-      const active = await countActiveWarnings(caller.platform, caller.userId);
-      if (active === 0) {
-        return text('You have no active warnings.');
-      }
-      if (active >= limit) {
-        return text(`You've reached the warning limit (${active}/${limit}). An admin can clear this.`);
-      }
-      let msg = `You have ${active} active warning${active === 1 ? '' : 's'} (limit ${limit}).`;
-      if (windowDays) {
-        const windowed = await countActiveWarnings(caller.platform, caller.userId, windowDays);
-        if (windowed < active) {
-          msg +=
-            ` ${active - windowed} of these are old enough not to count toward a new mute, but any uncleared ` +
-            'warning still applies if you leave and rejoin.';
-        }
-      }
-      return text(msg);
-    },
-  );
-
-  const appealModeration = tool(
-    'appeal_moderation',
-    "Ask the admins to review the caller's OWN active auto-moderation warning(s) — use when a member believes " +
-      'a warning (or being at/over the warning limit) was a false positive and wants a human to double-check. ' +
-      'NOT a general way to message admins — refuses cleanly with no active warnings (see suggest_improvement/' +
-      "report_content for other admin-notification paths). Always scoped to the caller's own platform/user id, " +
-      'never a model-supplied identifier — same self-scoping as my_warnings. Does not itself change any ' +
-      "warning or mute state — only an admin's clear_warnings can do that.",
-    {
-      reason: z
-        .string()
-        .max(APPEAL_MODERATION_REASON_MAX_CHARS)
-        .optional()
-        .describe(
-          "Optional short explanation of why the warning should be reviewed, in the member's own words " +
-            `(max ${APPEAL_MODERATION_REASON_MAX_CHARS} characters). Only pass through what they actually ` +
-            'said — never invent one.',
-        ),
-    },
-    async (args) => {
-      // Self-scoped, exactly like my_warnings: the eligibility gate reads
-      // ONLY caller.platform/caller.userId — there is no argument a model
-      // could supply to check or appeal on behalf of another user.
-      const active = await countActiveWarnings(caller.platform, caller.userId);
-      if (active === 0) {
-        return text("You don't currently have any active warnings to appeal.", true);
-      }
-      const cooldownHours = config.moderation.appealCooldownHours;
-      if (!reserveAppealSlot(`${caller.platform}:${caller.userId}`, cooldownHours)) {
-        return text(
-          `You've already asked for a review recently — please wait before appealing again ` +
-            `(once per ${cooldownHours}h).`,
-          true,
-        );
-      }
-      // Durable record FIRST (issue #554) — a missed/dismissed DM must never
-      // erase the appeal with no trace. Awaited, not fire-and-forget: the
-      // whole point of this write is that it survives even when the DM
-      // below fails, so it must actually land before we report success.
-      await createModerationAppeal({
-        platform: caller.platform,
-        userId: caller.userId,
-        userName: caller.userName,
-        reason: args.reason,
-        activeWarnings: active,
-        strikeLimit: config.moderation.strikeLimit,
-      });
-      void notifyAppealFiled(adapterFor, {
-        callerUserId: caller.userId,
-        callerName: caller.userName,
-        activeWarnings: active,
-        strikeLimit: config.moderation.strikeLimit,
-        reason: args.reason,
-      });
-      return text("Your appeal has been sent to the admins for review. They'll follow up if needed.");
-    },
-  );
-
-  const myData = tool(
-    'my_data',
-    'Summarize what the bot has stored about the caller: their own message count, replies the bot has ' +
-      'sent them, knowledge entries sourced from them, content reports and suggestions they filed, whether ' +
-      "they've published interests for member discovery, their standing response-style preference, and " +
-      "where they stand against today's daily reply budget. Use " +
-      'this when a member asks what the bot knows about them, wants to see what forget_me would erase ' +
-      'before deciding to invoke it, or asks how many messages they have left today. Read-only, scoped ' +
-      "exactly like forget_me — the caller's own identity plus any identity linked via link_member — so " +
-      "it can never see another member's data. Does not cover active warnings (see my_warnings) or the " +
-      'status of a specific filed item (see my_submissions), which already have their own tools; also ' +
-      'never includes admin notes about the caller (member_notes stays admin-only).',
-    {},
-    async () => {
-      const summary = await getMyDataSummary(caller.platform, caller.userId);
-      const lines = [
-        `Messages you've sent: ${summary.ownMessages}`,
-        `Replies the bot has sent you: ${summary.repliesToThem}`,
-        `Knowledge entries sourced from you: ${summary.knowledgeEntries}`,
-        `Content reports you've filed: ${summary.reportsFiled}`,
-        `Suggestions you've filed: ${summary.suggestionsFiled}`,
-        `Projects you've shared: ${summary.projectsShared}`,
-        `Interests published (who_is_into): ${summary.interestsPublished > 0 ? 'yes' : 'no'}`,
-        `Response style preference: ${summary.responseStyle === 'plain' ? 'plain' : 'standard (default)'}`,
-      ];
-      // Daily reply budget (issue #444) — reuses the exact function
-      // router.ts's own enforcement calls, so what this reports can never
-      // diverge from what actually gates the caller.
-      const limit = config.behaviour.dailyReplyLimitPerUser;
-      if (caller.role === 'super_admin') {
-        lines.push('Daily reply limit: exempt (super admin).');
-      } else if (limit === 0) {
-        lines.push('Daily reply limit: none configured.');
-      } else {
-        const used = await countRepliesToUser(caller.platform, caller.userId);
-        lines.push(
-          `Replies in the last 24h: ${used} / ${limit}` +
-            (used >= limit ? " — you've reached today's limit." : ''),
-        );
-      }
-      lines.push(
-        '',
-        'For your active warnings, use my_warnings. For the status of a specific report or suggestion, use my_submissions.',
-      );
-      return text(lines.join('\n'));
-    },
-  );
-
-  const suggestImprovement = tool(
-    'suggest_improvement',
-    "Record a member's suggestion for how this assistant/community bot could be improved, so the human " +
-      'maintainers see it. Capture only: a human reviews these and decides — never promise the change ' +
-      'will be built. The shared queue stays admin-only (triaged with list_suggestions); the member can ' +
-      'check their own status with my_submissions.',
-    {
-      content: z
-        .string()
-        .min(1)
-        .max(SUGGESTION_MAX_CHARS)
-        .describe(`The suggestion, in the member's own words (max ${SUGGESTION_MAX_CHARS} characters)`),
-    },
-    async (args) => {
-      const created = await createSuggestion({
-        platform: caller.platform,
-        userId: caller.userId,
-        displayName: caller.userName,
-        content: args.content,
-      });
-      if (!created) {
-        return text(
-          `You've already filed ${SUGGESTION_RATE_LIMIT_PER_DAY} suggestions in the last 24 hours. ` +
-            'Please wait before filing another.',
-          true,
-        );
-      }
-      return text(
-        `Suggestion #${created.id} recorded. A human maintainer reviews these — thanks for the idea, ` +
-          'but no promises on if/when it gets built.',
-      );
-    },
-  );
-
-  const suggestKnowledgeTool = tool(
-    'suggest_knowledge',
-    'Suggest a durable knowledge-base tip for other members — a hard-won answer, workaround, or fact ' +
-      'worth saving for the next person who hits the same thing (e.g. "for anyone else hitting this: X ' +
-      'needs Y"). This does NOT add to the knowledge base directly: it queues the tip in the SAME admin-' +
-      "reviewed candidate queue the offline context builder's own drafts use (list_knowledge_candidates) " +
-      '— nothing a member writes here can influence answers until an admin accepts it. Never call this ' +
-      'for a question, a request, or a feature/bot-improvement idea (use suggest_improvement for that).',
-    {
-      title: z
-        .string()
-        .min(1)
-        .max(KNOWLEDGE_TIP_TITLE_MAX_CHARS)
-        .describe(`Short FAQ-style title for the tip (max ${KNOWLEDGE_TIP_TITLE_MAX_CHARS} characters)`),
-      content: z
-        .string()
-        .min(1)
-        .max(KNOWLEDGE_TIP_CONTENT_MAX_CHARS)
-        .describe(
-          `The tip itself, in the member's own words (max ${KNOWLEDGE_TIP_CONTENT_MAX_CHARS} characters)`,
-        ),
-    },
-    async (args) => {
-      // SECURITY: tier is re-asserted here, not merely surface-gated by
-      // MEMBER_TOOLS — same defensive-double-check discipline every
-      // privileged/self-service tool in this file follows.
-      assertAtLeast(caller.role, 'member', 'suggest_knowledge');
-
-      // Topic = title, and this reuses the context builder's OWN pre-insert
-      // dedup guard verbatim (issue #503) so a member's tip is held to the
-      // same "don't refill an already-queued/reviewed or already-answered
-      // topic" bar as a machine-drafted candidate.
-      const topic = args.title;
-      const { blocked: alreadyQueued, embedding: topicEmbedding } =
-        await candidateTopicAlreadyReviewed(topic);
-      if (alreadyQueued) {
-        return text(
-          'Thanks, but a similar tip is already queued for review or has already been reviewed — no ' +
-            'need to resubmit.',
-        );
-      }
-      const covering = await findKnowledgeCoveringTopic(topicEmbedding);
-      if (covering) {
-        const label = covering.title ? `"${covering.title}"` : `entry #${covering.id}`;
-        return text(`Thanks, but this looks already covered by existing knowledge entry ${label}.`);
-      }
-
-      const created = await createKnowledgeTip({
-        platform: caller.platform,
-        userId: caller.userId,
-        topic,
-        title: args.title,
-        content: args.content,
-        topicEmbedding,
-      });
-      if (!created) {
-        return text(
-          `You've already suggested ${KNOWLEDGE_TIP_RATE_LIMIT_PER_DAY} tips in the last 24 hours. ` +
-            'Please wait before suggesting another.',
-          true,
-        );
-      }
-      return text(
-        `Thanks! Tip #${created.id} queued for admin review — it won't appear in the knowledge base ` +
-          'unless an admin accepts it.',
-      );
-    },
-  );
-
-  const rateAnswer = tool(
-    'rate_answer',
-    "Record whether the bot's most recent answer to the caller in this conversation was helpful. Call " +
-      'this ONLY on a clear, explicit member cue about the bot\'s own last answer (e.g. "that helped, ' +
-      'thanks", "that\'s wrong", a 👍/👎) — never on general positivity, ambiguous chatter, or feedback ' +
-      "about something other than the bot's last reply.",
-    {
-      helpful: z.boolean().describe('true if the answer helped, false if it did not'),
-      comment: z
-        .string()
-        .max(200)
-        .optional()
-        .describe(
-          "Optional short reason the member gave alongside the rating in the SAME message (e.g. 'wrong " +
-            "pricing, it changed last month'). Only pass through what they actually said — never invent one, " +
-            'and never ask a follow-up question just to solicit it.',
-        ),
-    },
-    async (args) => {
-      const created = await createAnswerFeedback({
-        platform: caller.platform,
-        conversationId: caller.conversationId,
-        userId: caller.userId,
-        helpful: args.helpful,
-        comment: args.comment,
-      });
-      if (created === 'no_recent_answer') {
-        return text("I don't have a recent answer of mine to rate in this conversation yet.", true);
-      }
-      if (created === 'rate_limited') {
-        return text(
-          `You've already rated ${RATE_ANSWER_DAILY_LIMIT} answers in the last 24 hours. ` +
-            'Please wait before rating another.',
-          true,
-        );
-      }
-      // Real-time admin escalation (issue #598): only a genuinely-recorded
-      // thumbs-down sets the turn-scoped flag — never a positive rating, and
-      // never the 'no_recent_answer'/'rate_limited' branches above, which
-      // return before reaching here. `notifyAdmins` is deliberately NOT
-      // called from this tool handler; `router.ts` reads this flag back
-      // post-turn and direct-fires it, preserving the documented "never from
-      // a model-callable tool" boundary (see `notifyAdmins`'s doc comment).
-      if (turnState && args.helpful === false) {
-        turnState.unhelpfulAnswerRated = true;
-      }
-      // Answered-question -> knowledge-base loop (issue #726,
-      // CAPABILITY-IDEAS.md §D2): a genuinely helpful, UNGROUNDED answer is
-      // silently drafted into the SAME admin-reviewed candidate queue
-      // suggest_knowledge (#633) writes into, via the exact same
-      // createKnowledgeTip/candidateTopicAlreadyReviewed/
-      // findKnowledgeCoveringTopic dedup+write path — the member-facing reply
-      // below is byte-identical either way. Fails closed whenever the
-      // grounding lookup can't recover a coherent preceding question (no
-      // `replyToUserId`, or no qualifying inbound row), never on the rater's
-      // own identity — the drafted row is attributed to the QUESTION's
-      // author (`grounding.questionUserId`), which can differ from the rater
-      // when `resolveAnswerFeedbackTarget` bound this rating to a reply
-      // addressed to someone else (SECURITY, issue #726 AC10).
-      // DM exclusion (issue #730 review): a 1:1 DM Q&A only ever enters the
-      // guild-wide, admin-visible candidate queue via the EXPLICIT
-      // suggest_knowledge act — never implicitly from a "helpful" rating.
-      // In a channel the exchange was already visible to the room; in a DM
-      // the member may reasonably assume privacy, and "helpful" is not
-      // consent to republish.
-      // Tier gate (issue #730 review, round 2): open-mode guests hold
-      // rate_answer (MEMBER_TOOLS surface), but writing into the
-      // knowledge_candidates queue is a member+ capability — suggest_knowledge
-      // asserts exactly that on the SAME createKnowledgeTip path. `atLeast`
-      // (not assertAtLeast) because the RATING itself stays guest-allowed:
-      // a guest's helpful rating records normally and only the drafting side
-      // effect is silently suppressed, same shape as the DM exclusion.
-      // Whole block is try/caught: drafting is a
-      // silent side effect on an already-recorded rating, so a transient
-      // failure in any of its reads/writes must degrade to "no draft" —
-      // never surface as a tool error on the rating itself (same fail-open
-      // posture as the gap/stale/retrieval supplements in knowledge_search).
-      if (
-        config.knowledgeAnswerCandidate.enabled &&
-        args.helpful === true &&
-        !caller.isDirect &&
-        atLeast(caller.role, 'member')
-      ) {
-        try {
-          const grounding = await answerFeedbackGrounding(created.interactionId);
-          if (
-            grounding &&
-            grounding.knowledgeEntryId === null &&
-            grounding.questionContent !== null &&
-            grounding.questionUserId !== null
-          ) {
-            // SECURITY (issue #726 follow-up): createKnowledgeTip's cap alone
-            // bounds how much of a single VICTIM's quota this can absorb, not
-            // how many DIFFERENT victims one rater can draft against via the
-            // mismatched-attribution fallback above — rate_answer's own daily
-            // cap (RATE_ANSWER_DAILY_LIMIT, 20/day) is far looser than any one
-            // victim's KNOWLEDGE_TIP_RATE_LIMIT_PER_DAY (3/day). A matched
-            // self-rating is exempt: that case is already bounded by
-            // createKnowledgeTip's own per-source-user cap. Fails closed
-            // (silently, same as every other branch here) rather than erroring
-            // the rating itself.
-            const mismatched = grounding.questionUserId !== caller.userId;
-            // Deliberately check-then-act, not atomic (issue #730 review): the
-            // count runs over answer_feedback AFTER this rating's own row was
-            // inserted, so the only overshoot window is calls truly in flight
-            // at the same instant, and the per-VICTIM cap inside
-            // createKnowledgeTip (an atomic INSERT..SELECT) stays the hard
-            // bound on actual damage regardless. A cross-table atomic rewrite
-            // would buy precision on an advisory secondary guard.
-            const raterExhausted =
-              mismatched &&
-              (await countMismatchedHelpfulRatings(caller.platform, caller.userId)) >
-                KNOWLEDGE_TIP_RATE_LIMIT_PER_DAY;
-            if (!raterExhausted) {
-              const { blocked, embedding: topicEmbedding } = await candidateTopicAlreadyReviewed(
-                grounding.questionContent,
-              );
-              if (!blocked) {
-                const covering = await findKnowledgeCoveringTopic(topicEmbedding);
-                if (!covering) {
-                  await createKnowledgeTip({
-                    platform: caller.platform,
-                    userId: grounding.questionUserId,
-                    topic: grounding.questionContent,
-                    title: grounding.questionContent,
-                    content: grounding.answerContent,
-                    topicEmbedding,
-                  });
-                }
-              }
-            }
-          }
-        } catch (err) {
-          logger.warn({ err }, 'rate_answer knowledge-candidate drafting failed; rating already recorded');
-        }
-      }
-      return text(args.helpful ? 'Thanks, glad that helped!' : 'Thanks for the feedback, noted.');
-    },
-  );
-
-  const requestHumanHelpTool = tool(
-    'request_human_help',
-    'Ask for a human community admin to be looped into this conversation. Call this ONLY on a clear, ' +
-      'explicit member ask for a human/admin (e.g. "can I talk to a human", "is there an admin I can ' +
-      'ask", "I need a person for this") — never on general frustration, ambiguous chatter, or a ' +
-      'question a normal answer can still address.',
-    {},
-    async () => {
-      // SECURITY: tier is re-asserted here, not merely surface-gated by
-      // MEMBER_TOOLS — same defensive-double-check discipline as
-      // suggest_knowledge/every other privileged/self-service tool in this
-      // file.
-      assertAtLeast(caller.role, 'member', 'request_human_help');
-
-      // Per-caller daily cap (issue #808): the ONE genuinely new piece this
-      // tool introduces — see reserveHumanHelpRequestSlot's doc comment for
-      // why. Checked before touching turnState, so a declined-by-cap call
-      // never sets the flag router.ts acts on.
-      const key = `${caller.platform}:${caller.userId}`;
-      if (!reserveHumanHelpRequestSlot(key, HUMAN_HELP_REQUEST_DAILY_LIMIT_PER_USER)) {
-        return text(
-          `You've already asked to talk to a human ${HUMAN_HELP_REQUEST_DAILY_LIMIT_PER_USER} times in ` +
-            'the last 24 hours. Please wait before asking again.',
-          true,
-        );
-      }
-
-      // Real-time admin escalation (issue #808): mirrors rate_answer's
-      // turn-scoped-flag discipline exactly (issue #598). `notifyAdmins` is
-      // deliberately NEVER called from this file — `router.ts` reads this
-      // flag back post-turn and direct-fires it from `msg.text`/
-      // `msg.userName`/`msg.platform`/`msg.conversationId` only, never from
-      // anything returned or accepted here (see `notifyAdmins`'s own doc
-      // comment). This is a zero-argument tool, so there is no model-composed
-      // free-text field that could reach that notification in the first
-      // place.
-      if (turnState) {
-        turnState.humanHelpRequested = true;
-      }
-      return text("Got it — I've flagged this for a community admin to follow up.");
-    },
-  );
-
-  const setResponseStyleTool = tool(
-    'set_response_style',
-    "Set the caller's standing reply style for every future message in every conversation, so they " +
-      "don't have to re-ask each time. Call with 'plain' when someone asks you to explain things more " +
-      'simply, avoid jargon, or use plainer language going forward — not for a one-off "explain that ' +
-      "again\" request, which should just be honoured directly in the reply. Call with 'standard' to " +
-      'revert to the normal style.',
-    { style: z.enum(['standard', 'plain']).describe('The reply style to use from now on') },
-    async (args) => {
-      await setResponseStyle(caller.platform, caller.userId, args.style);
-      return text(
-        args.style === 'plain'
-          ? "Got it — I'll keep replies simple and jargon-free from now on. Say the word to switch back."
-          : 'Got it — back to the normal reply style.',
-      );
-    },
-  );
-
-  const setLanguagePreferenceTool = tool(
-    'set_language_preference',
-    "Set the caller's standing reply language for every future message in every conversation, so " +
-      "they don't have to re-ask each time. Call with 'en' when someone asks you to always reply in " +
-      "NZ English from now on, or with 'mi' when someone asks you to always reply in te reo Māori " +
-      "from now on, regardless of what language their own messages are written in. Call with 'auto' " +
-      "to revert to today's default of mirroring whichever language their current message is in. Only " +
-      'call this for an explicit STANDING request ("always reply to me in Māori from now on") — a ' +
-      'one-off "reply in Māori just now" should just be honoured directly in the reply, without ' +
-      'calling this tool.',
-    { language: z.enum(['auto', 'en', 'mi']).describe('The reply language to use from now on') },
-    async (args) => {
-      await setLanguagePreference(caller.platform, caller.userId, args.language);
-      if (args.language === 'en') {
-        return text("Got it — I'll always reply in NZ English from now on. Say the word to switch back.");
-      }
-      if (args.language === 'mi') {
-        return text(
-          "Got it — I'll always reply in te reo Māori from now on where I can. Say the word to switch back.",
-        );
-      }
-      return text('Got it — back to mirroring whichever language you write in.');
-    },
-  );
-
-  const catchUp = tool(
-    'catch_up',
-    'Recap recent activity in the CURRENT conversation (this channel or DM) in chronological order, ' +
-      'for a member who has been away and wants to know what they missed. Always scoped to this ' +
-      'conversation only — call it with no arguments unless the member names a specific timeframe. ' +
-      'Use this for "what did I miss?", "what\'s been happening here?", "catch me up", and similar asks ' +
-      '— not for a topic-specific question, which remember_search answers better.',
-    {
-      hours: z
-        .number()
-        .positive()
-        .optional()
-        .describe(
-          `How many hours back to look (default ${CATCH_UP_DEFAULT_HOURS}). Hard-capped at ${CATCH_UP_MAX_HOURS} regardless of what's requested.`,
-        ),
-    },
-    async (args) => {
-      const hours = Math.min(args.hours ?? CATCH_UP_DEFAULT_HOURS, CATCH_UP_MAX_HOURS);
-      const since = new Date(Date.now() - hours * 60 * 60 * 1000);
-      // Always the caller's own real conversation — never a model-supplied
-      // id — identical scoping discipline to remember_search's default scope.
-      const entries = await recentConversationHistory(
-        caller.platform,
-        caller.conversationId,
-        since,
-        CATCH_UP_MAX_MESSAGES,
-      );
-      // Usage signal (issue #167 AC): a log counter of invocations plus the
-      // empty-vs-nonempty split, so adoption is measurable without a new
-      // table/migration.
-      logger.info(
-        {
-          platform: caller.platform,
-          conversationId: hashId(caller.conversationId),
-          hours,
-          resultCount: entries.length,
-        },
-        'catch_up invocation',
-      );
-      if (entries.length === 0) {
-        return text(`Nothing new here in the last ${hours} hour${hours === 1 ? '' : 's'}.`);
-      }
-      return text(
-        untrusted(
-          `Recent activity (last ${hours}h)`,
-          entries
-            .map((e) => {
-              const link = memoryHitJumpLink(e, config.discord.guildId);
-              // Same sanitization as remember_search above — the recalled
-              // author name is an untrusted, newline-unbounded display name
-              // (finding A).
-              const name = sanitizeName(e.userName);
-              return `[${e.createdAt.toISOString()}] [${e.direction}${name ? ` by ${name}` : ''}] ${e.content.slice(0, RECALL_TRUNCATION_CHARS)}${link ? ` (${link})` : ''}`;
-            })
-            .join('\n'),
-        ),
-      );
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const reactToMessage = tool(
-    'react_to_message',
-    'React to a message with an emoji instead of replying with text — a lightweight, low-noise ' +
-      `acknowledgement ("got it", "noted", "seen"). Only ${ALLOWED_REACTION_EMOJI.join(' ')} are allowed; ` +
-      'no other emoji, custom, or Nitro emoji can be used. Defaults to the message that triggered this ' +
-      'turn when messageId is omitted. Works on Discord and WhatsApp (both Baileys and Cloud API).',
-    {
-      emoji: z
-        .enum(ALLOWED_REACTION_EMOJI)
-        .describe(`One of: ${ALLOWED_REACTION_EMOJI.join(' ')} — no other value is accepted`),
-      messageId: z
-        .string()
-        .optional()
-        .describe('Message id to react to; defaults to the message that triggered this turn'),
-    },
-    async (args) => {
-      if (!adapter.reactToMessage) {
-        return text(`Reactions aren't available on ${caller.platform}.`, true);
-      }
-      const messageId = args.messageId ?? caller.messageId;
-      if (!messageId) {
-        return text('No message to react to — the current message has no visible id.', true);
-      }
-      // Same "the bot must have actually seen it" discipline as
-      // moderate/announce's target validation, scoped to the caller's own
-      // conversation (a member never names a different one).
-      if (!(await isKnownMessage(caller.platform, caller.conversationId, messageId))) {
-        return text(`Refusing: message "${messageId}" has never been seen in this conversation.`, true);
-      }
-      const key = `${caller.platform}:${caller.userId}`;
-      if (!reserveReactionDaily(key)) {
-        return text(
-          `You've hit today's reaction limit (${REACTION_RATE_LIMIT_PER_DAY}). Try again tomorrow.`,
-          true,
-        );
-      }
-      try {
-        await adapter.reactToMessage(caller.conversationId, messageId, args.emoji);
-        return text(`Reacted ${args.emoji}.`);
-      } catch (err) {
-        logger.warn({ err, actor: caller.userId }, 'react_to_message failed');
-        return text('Failed to react to that message.', true);
-      }
-    },
-    { annotations: { readOnlyHint: false } },
-  );
-
-  const setMyInterestsTool = tool(
-    'set_my_interests',
-    "Publish the caller's own interests — what they're building, learning, or into — so other members " +
-      `can find them via who_is_into (self-declared free text, max ${MEMBER_INTERESTS_MAX_CHARS} characters). ` +
-      "Call with the literal text 'clear' to remove the caller's entry and stop appearing in who_is_into " +
-      'results. Only call this on an explicit, deliberate request to publish/set/update/clear interests for ' +
-      'member discovery ("add me to who\'s into RAG", "update my interests", "clear my interests") — never ' +
-      'inferred from general chat about what someone is working on.',
-    {
-      interests: z
-        .string()
-        .min(1)
-        .max(MEMBER_INTERESTS_MAX_CHARS)
-        .describe(
-          "Free text describing the caller's own interests/what they're building or into (max " +
-            `${MEMBER_INTERESTS_MAX_CHARS} characters), or the literal string 'clear' to remove the caller's entry.`,
-        ),
-    },
-    async (args) => {
-      // Publishing to a member-facing discovery directory is a step further
-      // than a private, self-scoped preference like set_response_style, so
-      // this (and who_is_into below) explicitly floors at 'member' — same
-      // member-tier re-check share_project/list_projects below use.
-      assertAtLeast(caller.role, 'member', 'set_my_interests');
-      const { cleared } = await setMemberInterests(caller.platform, caller.userId, args.interests);
-      return text(
-        cleared
-          ? "Cleared your interests — you'll no longer appear in who_is_into results."
-          : 'Got it — your interests are now visible to other members via who_is_into.',
-      );
-    },
-  );
-
-  const whoIsIntoTool = tool(
-    'who_is_into',
-    'Find other members whose self-declared interests (published via set_my_interests) match a topic — ' +
-      'member-to-member discovery, e.g. "who\'s into RAG?" or "anyone working on MCP servers?". Returns up ' +
-      `to ${WHO_IS_INTO_LIMIT} matches by meaning. Results derive only from what members have explicitly ` +
-      'published with set_my_interests — never from general chat or any other source. A caller with no ' +
-      'published interests of their own can still search. Omit the topic to find members like the caller ' +
-      "themselves — matched against the caller's OWN published interests, excluding the caller's own entry " +
-      '(requires the caller to have already called set_my_interests).',
-    {
-      query: z
-        .string()
-        .min(1)
-        .max(300)
-        .optional()
-        .describe(
-          'Topic/keyword to search published member interests by meaning. Omit to search using the ' +
-            'caller\'s own published interests instead ("find people like me").',
-        ),
-    },
-    async (args) => {
-      assertAtLeast(caller.role, 'member', 'who_is_into');
-      if (args.query) {
-        const hits = await searchMemberInterests(args.query, WHO_IS_INTO_LIMIT);
-        if (hits.length === 0) {
-          return text('No members have published interests matching that yet.');
-        }
-        return text(await formatInterestResults(hits));
-      }
-      const selfMatch = await searchMemberInterestsForSelf(caller.platform, caller.userId, WHO_IS_INTO_LIMIT);
-      if (!selfMatch.hasProfile) {
-        // Issue #920: a caller with no published row of their own can no
-        // longer only be told to publish first — fall back to browsing the
-        // most recently published/updated interests (mirroring
-        // list_projects' no-query listRecentProjects default), still
-        // appending the same set_my_interests hint after the list.
-        const hint =
-          "You haven't published interests yet — call set_my_interests first, then who_is_into with no " +
-          'topic will search using your own published interests.';
-        const recent = await listRecentInterests(WHO_IS_INTO_LIMIT);
-        return text(recent.length === 0 ? hint : `${await formatInterestResults(recent)}\n\n${hint}`);
-      }
-      if (selfMatch.hits.length === 0) {
-        return text('No other members have published interests matching yours yet.');
-      }
-      return text(await formatInterestResults(selfMatch.hits));
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const setHelperAvailabilityTool = tool(
-    'set_helper_availability',
-    "Opt in or out of being notified when another member's find_helper topic matches the caller's own " +
-      'published interests (set via set_my_interests). Requires an existing published interests row — call ' +
-      "set_my_interests first if you haven't. Instantly reversible, same shape as set_response_style. Behind " +
-      'a feature flag; if unavailable on this server, this tool will not appear at all.',
-    {
-      available: z
-        .boolean()
-        .describe(
-          'true to opt in to being notified for find_helper requests matching your published interests; ' +
-            'false to opt out',
-        ),
-    },
-    async (args) => {
-      // MEMBER_TOOLS' floor re-check discipline for a self-service write that
-      // reaches other members, same as set_my_interests/share_project above.
-      assertAtLeast(caller.role, 'member', 'set_helper_availability');
-      if (!config.findHelper.enabled) {
-        return text('Peer-help handoff is not enabled on this server.', true);
-      }
-      const result = await setHelperAvailability(caller.platform, caller.userId, args.available);
-      if (!result.ok) {
-        return text(
-          "You don't have published interests yet — call set_my_interests first, then " +
-            'set_helper_availability can turn on notifications for topics matching them.',
-          true,
-        );
-      }
-      return text(
-        args.available
-          ? `You'll now be notified (at most ${FIND_HELPER_WEEKLY_LIMIT_PER_HELPER} times a week) when ` +
-              "another member's find_helper topic matches your published interests."
-          : "You won't be notified for find_helper requests anymore.",
-      );
-    },
-  );
-
-  const findHelperTool = tool(
-    'find_helper',
-    'Ask for member-to-member help on a topic. Matches your topic against members who have opted in via ' +
-      'set_helper_availability(true) and sends AT MOST ONE direct message, to the single best match — never ' +
-      'a broadcast. Never reveals who (if anyone) was contacted, their handle, or their interest text — only ' +
-      `whether someone was reached. Capped to ${FIND_HELPER_REQUESTER_DAILY_LIMIT} calls per rolling 24 hours. ` +
-      'Behind a feature flag; if unavailable on this server, this tool will not appear at all.',
-    {
-      topic: z
-        .string()
-        .min(1)
-        .max(FIND_HELPER_TOPIC_MAX_CHARS)
-        .describe(`What you need help with (max ${FIND_HELPER_TOPIC_MAX_CHARS} characters)`),
-    },
-    async (args) => {
-      assertAtLeast(caller.role, 'member', 'find_helper');
-      if (!config.findHelper.enabled) {
-        return text('Peer-help handoff is not enabled on this server.', true);
-      }
-      if (await isFindHelperRequesterAtDailyCap(caller.platform, caller.userId)) {
-        return text(
-          `You've hit today's ask-for-help limit (${FIND_HELPER_REQUESTER_DAILY_LIMIT}). Try again tomorrow.`,
-          true,
-        );
-      }
-      // Walks best-match-first; the FIRST candidate under its own weekly cap
-      // wins and the loop stops — at most one DM is ever sent per call
-      // (issue #729 SECURITY criterion), regardless of how many candidates
-      // matched. A candidate on a platform with no registered adapter in
-      // this deployment is skipped without consuming their weekly cap slot.
-      const candidates = await findHelperCandidates(args.topic, caller.platform, caller.userId);
-      for (const candidate of candidates) {
-        const target = adapterFor(candidate.platform);
-        if (!target) continue;
-        const claimed = await recordHelperNotificationIfUnderCap(
-          candidate.platform,
-          candidate.userId,
-          caller.platform,
-          caller.userId,
-          args.topic,
-        );
-        if (!claimed) continue;
-        const requesterLabel = await resolveSanitizedLabel(caller.platform, caller.userId);
-        // untrusted() quarantines the requester's free-text topic before it
-        // reaches a DIFFERENT member's DM (issue #729 SECURITY criterion) —
-        // same discipline list_answer_feedback's comment field already uses.
-        const message =
-          `${requesterLabel} could use some help with something you're into — reach out if you're able to.\n` +
-          untrusted('topic', args.topic);
-        // Best-effort send, same fire-and-forget/WindowClosedError-queue
-        // shape as notifySuggestionResolved/notifyReportResolved — a failed
-        // or queued send still counts as "the one DM this call sends" (the
-        // notification row above is already committed).
-        await target.sendDirectMessage(candidate.userId, message).catch((err) => {
-          if (err instanceof WindowClosedError && target.queueForWindowReopen) {
-            target.queueForWindowReopen(candidate.userId, message, 'low');
-            logger.warn(
-              { userId: hashId(candidate.userId), platform: candidate.platform },
-              "find_helper DM: recipient's window is closed, queued for reopen",
-            );
-            return;
-          }
-          logger.warn({ err, userId: hashId(candidate.userId) }, 'find_helper DM failed');
-        });
-        return text('Reached out to someone who may be able to help — hang tight.');
-      }
-      return text('No one available to help with that right now.');
-    },
-  );
-
-  const shareProjectTool = tool(
-    'share_project',
-    "Publish one of the caller's own projects to the community project showcase, visible to every " +
-      'other member via list_projects. Only call this on an explicit, deliberate request to share/' +
-      'showcase a project ("share my project", "add this to the showcase") — never inferred from ' +
-      'general chat about something someone is building. Calling with a name that matches one of the ' +
-      "caller's existing shared projects EDITS it in place (new description/link) rather than adding a " +
-      `duplicate. Capped at ${MEMBER_PROJECT_CAP} shared projects per member and ${PROJECT_RATE_LIMIT_PER_DAY} ` +
-      'new shares per rolling 24 hours — edits do not count against either cap. Set remove: true to take ' +
-      'an existing project down instead of adding or editing one.',
-    {
-      name: z
-        .string()
-        .min(1)
-        .max(PROJECT_NAME_MAX_CHARS)
-        .describe(
-          `The project's name (max ${PROJECT_NAME_MAX_CHARS} characters) — identifies which of the ` +
-            "caller's own projects this is, for edits and removal.",
-        ),
-      description: z
-        .string()
-        .min(1)
-        .max(PROJECT_DESCRIPTION_MAX_CHARS)
-        .optional()
-        .describe(
-          `What the project is/does, in the member's own words (max ${PROJECT_DESCRIPTION_MAX_CHARS} ` +
-            'characters). Required unless remove is true.',
-        ),
-      link: z
-        .string()
-        .max(PROJECT_LINK_MAX_CHARS)
-        .optional()
-        .describe(
-          'Optional URL to the project. Stored and shown to other members as plain text, verbatim — ' +
-            'the bot never fetches or previews it.',
-        ),
-      remove: z
-        .boolean()
-        .optional()
-        .describe(
-          "Set true to remove an existing project by name instead of adding/editing it — 'description' " +
-            "and 'link' are ignored when true.",
-        ),
-      seekingCollaborators: z
-        .boolean()
-        .optional()
-        .describe(
-          'Set true if the caller wants help/collaborators on this project — shown to other members via ' +
-            "list_projects. Defaults to false (showcase only). Only set this on the caller's own explicit " +
-            'statement, e.g. "I\'m looking for help with this" — never inferred.',
-        ),
-    },
-    async (args) => {
-      // Guests can reach every other MEMBER_TOOLS write in open mode, but
-      // publishing to a member-facing showcase is a step further than a
-      // self-scoped, invisible-to-others action like set_response_style —
-      // so this (and list_projects below) explicitly floor at 'member',
-      // the first MEMBER_TOOLS handler to do so (issue #646 AC #7).
-      assertAtLeast(caller.role, 'member', 'share_project');
-      if (args.remove) {
-        const removed = await removeMemberProject(caller.platform, caller.userId, args.name);
-        return removed
-          ? text(`Removed "${args.name}" from the project showcase.`)
-          : text(`You don't have a shared project named "${args.name}".`, true);
-      }
-      if (!args.description) {
-        return text('A description is required to share or edit a project.', true);
-      }
-      const result = await shareProject({
-        platform: caller.platform,
-        userId: caller.userId,
-        name: args.name,
-        description: args.description,
-        link: args.link,
-        seekingCollaborators: args.seekingCollaborators,
-      });
-      if (!result.ok) {
-        return text(
-          result.reason === 'cap'
-            ? `You already have ${MEMBER_PROJECT_CAP} shared projects — remove one first (share_project ` +
-                'with remove: true) before adding another.'
-            : `You've already shared ${PROJECT_RATE_LIMIT_PER_DAY} new projects in the last 24 hours. ` +
-                'Please wait before sharing another.',
-          true,
-        );
-      }
-      return text(
-        result.created
-          ? `Shared "${args.name}" — other members can find it with list_projects.`
-          : `Updated "${args.name}".`,
-      );
-    },
-  );
-
-  const listProjectsTool = tool(
-    'list_projects',
-    'Browse the member-declared project showcase — what other members have built and published with ' +
-      'share_project. With no query, returns the most recently shared projects; with a query, returns ' +
-      'the closest matches by meaning (e.g. "anyone working on a Discord bot?" or "RAG projects"). ' +
-      'Results derive only from what members have explicitly shared — never from general chat. Links ' +
-      'render as plain text and are never fetched.',
-    {
-      query: z
-        .string()
-        .max(300)
-        .optional()
-        .describe(
-          'Optional topic/keyword to search shared projects by meaning. Omit for the most recently ' +
-            'shared projects.',
-        ),
-      seekingCollaborators: z
-        .boolean()
-        .optional()
-        .describe('Only show projects whose owner is looking for collaborators. Omit or false to show all.'),
-      mine: z
-        .boolean()
-        .optional()
-        .describe(
-          "Only show the caller's own shared projects — ignores query/seekingCollaborators when set. " +
-            'Use this to find the exact name of one of your own projects before editing or removing it ' +
-            'with share_project.',
-        ),
-    },
-    async (args) => {
-      assertAtLeast(caller.role, 'member', 'list_projects');
-      if (args.mine) {
-        const projects = await listOwnProjects(caller.platform, caller.userId);
-        if (projects.length === 0) {
-          return text("You haven't shared any projects yet.");
-        }
-        return text(await formatProjectResults(projects));
-      }
-      const opts = { seekingCollaboratorsOnly: args.seekingCollaborators };
-      const projects = args.query
-        ? await searchProjects(args.query, LIST_PROJECTS_DEFAULT_LIMIT, opts)
-        : await listRecentProjects(LIST_PROJECTS_DEFAULT_LIMIT, opts);
-      if (projects.length === 0) {
-        return text(
-          args.seekingCollaborators
-            ? 'No projects are currently looking for collaborators.'
-            : args.query
-              ? 'No shared projects match that.'
-              : 'No projects have been shared yet.',
-        );
-      }
-      return text(await formatProjectResults(projects));
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const requestProjectConnectionTool = tool(
-    'request_project_connection',
-    'Ask to connect with a project owner who marked their project 🤝 looking for collaborators (via ' +
-      "share_project's seekingCollaborators) — the action counterpart to that signal. Looks up the project " +
-      'by the id shown in list_projects/who_is_into (e.g. 42 from "[#42]") and sends the owner AT MOST ONE ' +
-      "direct message naming the caller and the project — never a broadcast, and never discloses the owner's " +
-      `identity/handle back to the caller beyond what list_projects already showed. Capped to ` +
-      `${PROJECT_CONNECTION_REQUESTER_DAILY_LIMIT} calls per rolling 24 hours. Refuses cleanly if the project ` +
-      "isn't found, isn't seeking collaborators, or belongs to the caller.",
-    {
-      projectId: z
-        .number()
-        .int()
-        .positive()
-        .describe(
-          'The id of the project to request a connection for, as shown by list_projects (e.g. 42 from "[#42]").',
-        ),
-    },
-    async (args) => {
-      // Publishing to a member-facing showcase (share_project/list_projects)
-      // floors at 'member', excluding open-mode guests — this reaches a
-      // DIFFERENT member's DM, so it inherits the same floor.
-      assertAtLeast(caller.role, 'member', 'request_project_connection');
-      // Requester daily-cap check FIRST, before any project lookup — same
-      // order-of-operations as find_helper's isFindHelperRequesterAtDailyCap
-      // check (issue #729 AC #6 precedent).
-      if (await isProjectConnectionRequesterAtDailyCap(caller.platform, caller.userId)) {
-        return text(
-          `You've hit today's connection-request limit (${PROJECT_CONNECTION_REQUESTER_DAILY_LIMIT}). ` +
-            'Try again tomorrow.',
-          true,
-        );
-      }
-      const project = await getActiveProjectById(args.projectId);
-      if (!project) {
-        return text('No active project with that id.', true);
-      }
-      if (!project.seekingCollaborators) {
-        return text('That project is not currently looking for collaborators.', true);
-      }
-      // Self-match structurally impossible: this check runs BEFORE any DB
-      // write (issue #729's find_helper precedent for self-exclusion).
-      if (project.platform === caller.platform && project.userId === caller.userId) {
-        return text("You can't request to connect with your own project.", true);
-      }
-      const target = adapterFor(project.platform);
-      if (!target) {
-        return text("That project's owner can't be reached on this deployment right now.", true);
-      }
-      const claimed = await recordProjectConnectionIfUnderCap(
-        project.platform,
-        project.userId,
-        caller.platform,
-        caller.userId,
-        project.id,
-      );
-      if (!claimed) {
-        // Generic refusal — never discloses the owner's cap state, same
-        // discipline as find_helper's "no one available" message.
-        return text("That project's owner can't receive new connection requests right now.", true);
-      }
-      const requesterLabel = await resolveSanitizedLabel(caller.platform, caller.userId);
-      // untrusted() quarantines the member-supplied project name before it
-      // reaches a DIFFERENT member's DM (issue #840 SECURITY criterion) —
-      // same discipline find_helper's topic field already uses.
-      const message =
-        `${requesterLabel} is interested in collaborating on ` +
-        `${untrusted('project', project.name)} — reach out if you're able to.`;
-      // Best-effort send, same fire-and-forget/WindowClosedError-queue shape
-      // as find_helper/notifySuggestionResolved — a failed or queued send
-      // still counts as "the one DM this call sends" (the connection-request
-      // row above is already committed).
-      await target.sendDirectMessage(project.userId, message).catch((err) => {
-        if (err instanceof WindowClosedError && target.queueForWindowReopen) {
-          target.queueForWindowReopen(project.userId, message, 'low');
-          logger.warn(
-            { userId: hashId(project.userId), platform: project.platform },
-            "request_project_connection DM: recipient's window is closed, queued for reopen",
-          );
-          return;
-        }
-        logger.warn({ err, userId: hashId(project.userId) }, 'request_project_connection DM failed');
-      });
-      return text('Reached out to the project owner — hang tight.');
-    },
-  );
-
-  const communityDigestTool = tool(
-    'community_digest',
-    "On-demand pull of the community digest — the same this-week's-topics, new-in-the-knowledge-base, " +
-      'project-showcase, and platform-update signals the weekly member digest post would send right now, ' +
-      'without waiting for its cadence. Takes no arguments; read-only; does not affect when the next ' +
-      'scheduled weekly digest post goes out. Member only.',
-    {},
-    async () => {
-      // Same MEMBER_TOOLS floor re-check discipline as who_is_into/
-      // share_project — publishes aggregate community-wide content, so this
-      // excludes open-mode guests even though the tool is structurally
-      // reachable at MEMBER_TOOLS tier.
-      assertAtLeast(caller.role, 'member', 'community_digest');
-      const message = await buildMemberDigestContent();
-      if (message == null) return text('Nothing to report right now.');
-      // This tool result re-enters the model's context (unlike the weekly
-      // channel post, sent straight to Discord) — quarantine it the same way
-      // admin_digest quarantines buildAdminDigestForAdmin's own return
-      // (issue #499 review), since this text embeds model-cluster-summarized
-      // topic labels, not a direct human DM.
-      return text(untrusted('Community digest', message));
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  // --- Project tools (issue #927) --------------------------------------------
-  //
-  // Member tier, like every other tool in this section. Being in a project is
-  // DATA SCOPE, NOT A TIER: these tools are on every member's surface and are
-  // simply inert for someone with no visible project, so nothing here changes
-  // what `toolsForRole` derives. Both access checks (membership, expanded
-  // through linked identities; and surface, i.e. a bound conversation or a DM)
-  // live in SQL in `visibleProjectIds` — never re-derived here.
-
-  const projectRecall = tool(
-    'project_recall',
-    'Search the shared memory of a project you are part of — decisions, notes and references the team ' +
-      'saved. Use this whenever someone asks what the team decided, agreed, or recorded about something. ' +
-      'Only ever returns content from projects you are a member of, and only in a conversation that ' +
-      'project is bound to.',
-    { query: z.string().describe('What to look up in the project memory') },
-    async (args) => {
-      // SECURITY: re-check member tier in the handler, the same discipline
-      // share_project/set_my_interests/who_is_into/find_helper/community_digest
-      // already use (see rbac.ts). MEMBER_TOOLS is also a GUEST's surface in
-      // open mode ("Guests only ever reach the agent in open mode; same
-      // surface as member"), and visibleProjectIds intentionally checks only
-      // project_members — never tier — so without this an open-mode guest who
-      // still has a membership row reads a team's private notes. That is the
-      // mechanism behind the removed-member leak fixed in removeMember too
-      // (PR #929 review).
-      assertAtLeast(caller.role, 'member', 'project_recall');
-      const hits = await searchProjectNotes(args.query, {
-        platform: caller.platform,
-        userId: caller.userId,
-        conversationId: caller.conversationId,
-        isDirect: caller.isDirect,
-      });
-      if (hits.length === 0) {
-        return text('Nothing in project memory matches that (or you have no project accessible here).');
-      }
-      recordProjectNoteRetrieval(hits.map((h) => h.id)).catch((err) =>
-        logger.warn({ err }, 'Project note retrieval count update failed'),
-      );
-      // Notes are member-authored free text re-entering the model's context,
-      // so they are quarantined exactly as community_digest and admin_digest
-      // quarantine theirs — context, never instructions.
-      return text(
-        untrusted(
-          'Project memory',
-          hits
-            .map((h) => {
-              const ref = h.referenceUrl ? `\n  reference: ${h.referenceUrl}` : '';
-              return `- [${h.projectSlug}] ${h.title ? `${h.title}: ` : ''}${h.content}${ref}`;
-            })
-            .join('\n'),
-        ),
-      );
-    },
-    { annotations: { readOnlyHint: true } },
-  );
-
-  const projectNote = tool(
-    'project_note',
-    'Record a decision, note or document reference in a project you are part of, so the team can find ' +
-      'it later. Use this when someone says to remember/record/note something for the project. The ' +
-      'reference link is stored verbatim and never opened.',
-    {
-      project: z.string().describe('The project slug (see project_list)'),
-      content: z
-        .string()
-        .min(1)
-        .max(PROJECT_NOTE_CONTENT_MAX_CHARS)
-        .describe(`What to record (max ${PROJECT_NOTE_CONTENT_MAX_CHARS} characters)`),
-      title: z
-        .string()
-        .max(PROJECT_NOTE_TITLE_MAX_CHARS)
-        .optional()
-        .describe(`Short label for the note (max ${PROJECT_NOTE_TITLE_MAX_CHARS} characters)`),
-      referenceUrl: z
-        .string()
-        .url()
-        .max(PROJECT_NOTE_REFERENCE_URL_MAX_CHARS)
-        .optional()
-        .describe('Optional link to an external doc — stored, never fetched'),
-    },
-    async (args) => {
-      // SECURITY: re-check member tier in the handler, the same discipline
-      // share_project/set_my_interests/who_is_into/find_helper/community_digest
-      // already use (see rbac.ts). MEMBER_TOOLS is also a GUEST's surface in
-      // open mode ("Guests only ever reach the agent in open mode; same
-      // surface as member"), and visibleProjectIds intentionally checks only
-      // project_members — never tier — so without this an open-mode guest who
-      // still has a membership row reads a team's private notes. That is the
-      // mechanism behind the removed-member leak fixed in removeMember too
-      // (PR #929 review).
-      assertAtLeast(caller.role, 'member', 'project_note');
-      const saved = await saveProjectNote(
-        {
-          platform: caller.platform,
-          userId: caller.userId,
-          conversationId: caller.conversationId,
-          isDirect: caller.isDirect,
-        },
-        {
-          slug: args.project,
-          content: args.content,
-          title: args.title,
-          referenceUrl: args.referenceUrl,
-        },
-      );
-      // Deliberately the same reply for "no such project" and "exists but not
-      // yours / not bound here" (issue #205's wording rule): distinguishing
-      // them would confirm a project's existence to a non-member.
-      if (!saved) return text('No project by that name is accessible here.', true);
-      // A rolling-24h write cap, same refusal shape as suggest_knowledge's
-      // (PR #929 review). Deliberately worded as a limit that resets, not as
-      // a rejection of the content, so a team minuting a long meeting knows
-      // the note simply needs to wait rather than being lost to a bug.
-      if ('atCap' in saved) {
-        return text(
-          `You've already recorded ${PROJECT_NOTE_RATE_LIMIT_PER_DAY} project notes in the last 24 ` +
-            'hours. Try again later, or ask an admin if the team needs a higher limit.',
-          true,
-        );
-      }
-      return text(`Recorded in ${args.project}.`);
-    },
-    { annotations: { readOnlyHint: false } },
-  );
-
-  const projectList = tool(
-    'project_list',
-    'List the projects you can access in this conversation, with their standing brief. Use this when ' +
-      'someone asks what projects they are in or what a project is about.',
-    {},
-    async () => {
-      // SECURITY: re-check member tier in the handler, the same discipline
-      // share_project/set_my_interests/who_is_into/find_helper/community_digest
-      // already use (see rbac.ts). MEMBER_TOOLS is also a GUEST's surface in
-      // open mode ("Guests only ever reach the agent in open mode; same
-      // surface as member"), and visibleProjectIds intentionally checks only
-      // project_members — never tier — so without this an open-mode guest who
-      // still has a membership row reads a team's private notes. That is the
-      // mechanism behind the removed-member leak fixed in removeMember too
-      // (PR #929 review).
-      assertAtLeast(caller.role, 'member', 'project_list');
-      const projects = await listVisibleProjects({
-        platform: caller.platform,
-        userId: caller.userId,
-        conversationId: caller.conversationId,
-        isDirect: caller.isDirect,
-      });
-      if (projects.length === 0) return text('You have no project accessible in this conversation.');
-      return text(
-        untrusted(
-          'Projects',
-          projects.map((p) => `- ${p.name} [${p.slug}]${p.brief ? `\n  ${p.brief}` : ''}`).join('\n'),
-        ),
-      );
-    },
-    { annotations: { readOnlyHint: true } },
   );
 
   // --- Admin tools (scoped to the admin's own conversations) ------------------
@@ -5655,14 +3613,6 @@ export function buildToolServer(
     version: '2.0.0',
     tools: [
       ...registryTools,
-      communityInfo,
-      communityGuidelines,
-      checkStatus,
-      listEvents,
-      knowledgeSearch,
-      projectRecall,
-      projectNote,
-      projectList,
       projectCreate,
       projectAddMember,
       projectRemoveMember,
@@ -5671,32 +3621,6 @@ export function buildToolServer(
       projectInfo,
       projectArchive,
       projectUnarchive,
-      listKnowledgeTopicsTool,
-      rememberSearch,
-      forgetMe,
-      reportContent,
-      withdrawReport,
-      withdrawKnowledgeTip,
-      mySubmissions,
-      myWarnings,
-      appealModeration,
-      myData,
-      suggestImprovement,
-      suggestKnowledgeTool,
-      rateAnswer,
-      requestHumanHelpTool,
-      setResponseStyleTool,
-      setLanguagePreferenceTool,
-      catchUp,
-      reactToMessage,
-      setMyInterestsTool,
-      whoIsIntoTool,
-      setHelperAvailabilityTool,
-      findHelperTool,
-      shareProjectTool,
-      listProjectsTool,
-      requestProjectConnectionTool,
-      communityDigestTool,
       whatsNew,
       userHistory,
       moderate,
